% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{addArrows}
\alias{addArrows}
\title{Add covariates to arrows}
\usage{
addArrows(arrows, covar)
}
\arguments{
\item{arrows}{A data.frame}

\item{covar}{A list of covariates}
}
\description{
Add covariates to arrows
}
