% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catInteraction.R
\name{addCatVar}
\alias{addCatVar}
\title{Add dummy vars to data.frame}
\usage{
addCatVar(df, varname, groupLetter = "d", mode = 1)
}
\arguments{
\item{df}{A data.frame}

\item{varname}{Variable name to be converted as factor and add dummies}

\item{groupLetter}{A character}

\item{mode}{if mode is 2, apply different coding system}
}
\description{
Add dummy vars to data.frame
}
\examples{
addCatVar(mtcars,"cyl")
protest1=addCatVar(protest,"protest",mode=2)
head(protest1)
}
