% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2table.R
\name{addLabels}
\alias{addLabels}
\title{add name to labels}
\usage{
addLabels(labels, id, name)
}
\arguments{
\item{labels}{A list}

\item{id}{label id}

\item{name}{A character}
}
\description{
add name to labels
}
\examples{
labels=c(X="X",M="M",Y="Y")
addLabels(labels,"W","X")
addLabels(labels,"W","W")
}
