% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{addLatentNodes}
\alias{addLatentNodes}
\title{Add latent nodes information to nodes}
\usage{
addLatentNodes(nodes, fit, labels)
}
\arguments{
\item{nodes}{A data.frame}

\item{fit}{An object of class lavaan. Result of lavaan::sem()}

\item{labels}{A list}
}
\description{
Add latent nodes information to nodes
}
