% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{addLine}
\alias{addLine}
\title{Add line feed to string}
\usage{
addLine(x, ...)
}
\arguments{
\item{x}{A string}

\item{...}{one or more R objects, to be converted to character vectors.}
}
\description{
Add line feed to string
}
