% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{addNodes}
\alias{addNodes}
\title{Add covariates to nodes}
\usage{
addNodes(nodes, covar, radx = 0.1, rady = 0.04, no = NULL)
}
\arguments{
\item{nodes}{A data.frame}

\item{covar}{A list of covariates}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{no}{A numeric}
}
\description{
Add covariates to nodes
}
