% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{addPlus}
\alias{addPlus}
\title{Add `+` mark to string}
\usage{
addPlus(x, ...)
}
\arguments{
\item{x}{A string}

\item{...}{one or more R objects, to be converted to character vectors.}
}
\description{
Add `+` mark to string
}
