% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCatEquation.R
\name{addTripleInteraction}
\alias{addTripleInteraction}
\title{Add triple interaction}
\usage{
addTripleInteraction(res, names, interactionNo = 0)
}
\arguments{
\item{res}{A character vector}

\item{names}{A character vector}

\item{interactionNo}{A numeric}
}
\description{
Add triple interaction
}
