% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2table.R
\name{appendLabels}
\alias{appendLabels}
\title{Append labels from vars, moderator and covar}
\usage{
appendLabels(labels, vars = list(), moderator = list(), covar = NULL)
}
\arguments{
\item{labels}{A list}

\item{vars}{A list}

\item{moderator}{A list}

\item{covar}{A list}
}
\description{
Append labels from vars, moderator and covar
}
