% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationBK.R
\name{bda.mediation.test}
\alias{bda.mediation.test}
\title{The Sobel mediation test}
\usage{
bda.mediation.test(mv, iv, dv)
}
\arguments{
\item{mv}{The mediator variable}

\item{iv}{The independent variable}

\item{dv}{The dependent variable}
}
\description{
To compute statistics and p-values for the Sobel test. Results for three versions of "Sobel test" are provided: Sobel test, Aroian test and Goodman test.
}
