% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{caskets}
\alias{caskets}
\title{CASKETS dataset}
\format{
A data.frame with 541 obs. of  7 variables
\describe{
\item{policy}{Given information about policy (0 = No information, 1 = Told About Policy)}
\item{interest}{Interest in viewing casket images}
\item{age}{Participant age}
\item{educ}{Participant education level, 1 = less than high school, 2 = high school, 3 = some college, 4 = associates or technical school, 5 = bachelor degree, 6 = some graduate school, 7 = graduate degree }
\item{male}{Participant sex (0 = female, 1 = male)}
\item{conserv}{Participant social conservatism}
\item{kerry}{Kerry or Bush supporter, 0 = bush supporter, 1 = kerry supporter}
}
}
\source{
Hayes, A. F., & Reineke, J. B. (2007). The effects of government censorship of war-related news coverage on interest in the censored coverage: A test of competing theories. Mass Communication and Society, 10, 423-438

\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
caskets
}
\description{
CASKETS dataset
}
\keyword{datasets}
