% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMediation.R
\name{catEquation}
\alias{catEquation}
\title{Make Equation with categorical variable}
\usage{
catEquation(X = "A", Y = "B", cat = NULL, W = NULL, count = NULL,
  data = NULL, prefix = "c", maxylev = 6, groupLetter = c("d",
  "e"))
}
\arguments{
\item{X}{Name of independent variable}

\item{Y}{Name of dependent variable}

\item{cat}{Which variable is categorical? Choices are c("X","Y")}

\item{W}{Names of moderator variables}

\item{count}{length of unique values of categorical variable}

\item{data}{A data.frame}

\item{prefix}{A prefix}

\item{maxylev}{maximal unique length of categorial variable}

\item{groupLetter}{A character vector}
}
\description{
Make Equation with categorical variable
}
\examples{
catEquation(count=4)
catEquation(cat="Y",count=4,prefix="b")
catEquation(W="W",count=4)
catEquation(cat="X",W="W",count=4)
catEquation(W="W",cat="Y",count=4)
catEquation(X="cyl",Y="wt",data=mtcars)
catEquation(X="wt",Y="cyl",data=mtcars)
catEquation(X="cyl",Y="am",W="wt",data=mtcars)
}
