% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catInteraction.R
\name{catInteraction}
\alias{catInteraction}
\title{Make interaction equation with dummy categorical and/or moderator}
\usage{
catInteraction(Y, X, W, data = NULL, prefix = "b", covar = list(),
  corr = TRUE, maxylev = 6)
}
\arguments{
\item{Y}{Name of dependent variable}

\item{X}{Name of independent variable}

\item{W}{Name of moderator variable}

\item{data}{A data.frame}

\item{prefix}{A prefix}

\item{covar}{A list}

\item{corr}{A logical}

\item{maxylev}{maximal unique length of categorial variable}
}
\description{
Make interaction equation with dummy categorical and/or moderator
}
\examples{
cat(catInteraction(Y="mpg",X="wt",W="cyl",data=mtcars))
cat(catInteraction(Y="mpg",X="carb",W="cyl",data=mtcars))
cat(catInteraction(Y="mpg",X="wt",W="hp",data=mtcars))
cat(catInteraction(Y="mpg",X="cyl",W="hp",data=mtcars))
}
