% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catInteraction.R
\name{catInteraction2}
\alias{catInteraction2}
\title{Make interaction equation with dummy categorical moderator}
\usage{
catInteraction2(Y, X, W = NULL, data = NULL, count = NULL,
  prefix = "b", covar = list(), corr = TRUE)
}
\arguments{
\item{Y}{Name of dependent variable}

\item{X}{Optional.Name of independent variable}

\item{W}{Name of moderator variable}

\item{data}{A data.frame}

\item{count}{length of unique values of independent variable}

\item{prefix}{A prefix}

\item{covar}{A list}

\item{corr}{A logical}
}
\description{
Make interaction equation with dummy categorical moderator
}
\examples{
cat(catInteraction2(Y="mpg",X="wt",count=3))
cat(catInteraction2(Y="mpg",X="wt",W="cyl",data=mtcars))
}
