% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catInteraction.R
\name{catInteraction3}
\alias{catInteraction3}
\title{Make interaction equation with dummy categorical antecedent and moderator}
\usage{
catInteraction3(Y, X = NULL, W = NULL, data = NULL, xcount = NULL,
  wcount = NULL, prefix = "b", covar = list())
}
\arguments{
\item{Y}{Name of dependent variable}

\item{X}{Name of independent variable}

\item{W}{Optional.Name of moderator variable}

\item{data}{A data.frame}

\item{xcount}{length of unique values of independent variable}

\item{wcount}{length of unique values of moderator variable}

\item{prefix}{A prefix}

\item{covar}{A list}
}
\description{
Make interaction equation with dummy categorical antecedent and moderator
}
\examples{
cat(catInteraction2(Y="mpg",X="carb",W="cyl",data=mtcars))
}
