% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMediation.R
\name{catMediation}
\alias{catMediation}
\title{Make Mediation Equation with one categorical variable}
\usage{
catMediation(X = "X", M = "M", Y = "Y", cat = NULL, count = NULL,
  data = NULL, moderator = list(), covar = NULL, mode = 0,
  maxylev = 6, range = TRUE, rangemode = 1)
}
\arguments{
\item{X}{Name of independent variable}

\item{M}{Name of mediator variable}

\item{Y}{Name of dependent variable}

\item{cat}{Which variable is categorical? Choices are c("X","Y")}

\item{count}{length of unique values of categorical variable}

\item{data}{A data.frame}

\item{moderator}{A list}

\item{covar}{A list of covariates}

\item{mode}{A numeric. 0: SEM equation, 1: regression equation}

\item{maxylev}{maximal unique length of categorial variable}

\item{range}{A logical}

\item{rangemode}{range mode}
}
\description{
Make Mediation Equation with one categorical variable
}
\examples{
moderator=list(name=c("wt"),site=list(c("a","c")))
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M","Y"),"Y","Y"))
cat(catMediation(cat="X",count=4))
cat(catMediation(cat="X",count=4,covar=covar))
cat(catMediation(cat="M",count=4))
cat(catMediation(cat="Y",count=4))
cat(catMediation(cat="X",count=3,moderator=moderator))
cat(catMediation(cat="M",count=4,moderator=moderator))
cat(catMediation(cat="Y",count=4,moderator=moderator))
mtcars=addCatVar(mtcars,"cyl")
cat(catMediation(X="cyl",M="am",Y="mpg",data=mtcars,moderator=moderator))
}
