% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{centerPrint}
\alias{centerPrint}
\title{Print a string in center}
\usage{
centerPrint(string, width)
}
\arguments{
\item{string}{A string}

\item{width}{A numeric}
}
\description{
Print a string in center
}
