% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleMediation.R
\name{checkEquationVars}
\alias{checkEquationVars}
\title{Check dependent variables in equations}
\usage{
checkEquationVars(equation)
}
\arguments{
\item{equation}{A string of regression formula}
}
\description{
Check dependent variables in equations
}
\examples{
equation="M1~X*M*W+W*Z\nM2~X+M1+X"
checkEquationVars(equation)
}
