% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleMediation.R
\name{checkEqVars}
\alias{checkEqVars}
\title{Check dependent variables in a equation}
\usage{
checkEqVars(eq)
}
\arguments{
\item{eq}{A string of regression formula}
}
\description{
Check dependent variables in a equation
}
\examples{
eq="M2~X+M+X+X*M*W"
checkEqVars(eq)
}
