% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{compareMC}
\alias{compareMC}
\title{Compare effects of mean-centering and standardization of model}
\usage{
compareMC(fit, mode = 1)
}
\arguments{
\item{fit}{An object of class 'lm'}

\item{mode}{integer}
}
\value{
if mode is 1, an object of modelSummary2. Otherwise lift of models
}
\description{
Compare effects of mean-centering and standardization of model
}
\examples{
fit=lm(govact~negemot*age,data=glbwarm)
compareMC(fit)
compareMC(fit,mode=2)
}
