% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{compareMCTable}
\alias{compareMCTable}
\title{Make table comparing effects of mean-centering and standardization of model}
\usage{
compareMCTable(fit, vanilla = TRUE)
}
\arguments{
\item{fit}{An object of class 'lm'}

\item{vanilla}{logical.}
}
\description{
Make table comparing effects of mean-centering and standardization of model
}
