% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{compareVIF}
\alias{compareVIF}
\title{Compare correlation, tolerance, vif of mean-centered and standardized models}
\usage{
compareVIF(fit)
}
\arguments{
\item{fit}{An object of class lm}
}
\description{
Compare correlation, tolerance, vif of mean-centered and standardized models
}
\examples{
fit=lm(govact~negemot*age,data=glbwarm)
compareVIF(fit)
}
