% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{compareVIFTable}
\alias{compareVIFTable}
\title{Make table comparing correlation, tolerance, vif of mean-centered and standardized models}
\usage{
compareVIFTable(fit, vanilla = TRUE)
}
\arguments{
\item{fit}{An object of class lm}

\item{vanilla}{logical}
}
\description{
Make table comparing correlation, tolerance, vif of mean-centered and standardized models
}
\examples{
fit=lm(govact~negemot*age,data=glbwarm)
compareVIFTable(fit)
compareVIFTable(fit,vanilla=FALSE)
}
