% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{conceptDiagram}
\alias{conceptDiagram}
\title{Make conceptDiagram}
\usage{
conceptDiagram(fit, labels = NULL)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{labels}{labels}
}
\description{
Make conceptDiagram
}
