% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalEffectPlot.R
\name{conditionalEffectPlot}
\alias{conditionalEffectPlot}
\title{Make conditional effect plot}
\usage{
conditionalEffectPlot(fit, values = NULL, data, mod = "skeptic")
}
\arguments{
\item{fit}{An object of class lavaan}

\item{values}{Optional value}

\item{data}{A data.frame}

\item{mod}{Name of moderator variable}
}
\description{
Make conditional effect plot
}
