% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlot.R
\name{condPlot}
\alias{condPlot}
\title{Draw conditional effect plot}
\usage{
condPlot(fit, pred = NULL, modx = NULL, pred.values = NULL,
  modx.values = NULL, mode = 1, rangemode = 1, hjust = NULL,
  digits = 3, ...)
}
\arguments{
\item{fit}{An onject of class lm}

\item{pred}{name of predictor variable}

\item{modx}{name of moderator variable}

\item{pred.values}{Values of predictor variables}

\item{modx.values}{Values of modifier variables}

\item{mode}{integer. one of 1:3.}

\item{rangemode}{integer. 1 or 2}

\item{hjust}{hjust of label}

\item{digits}{integer indicating the number of decimal places}

\item{...}{further arguments to be passed to add_lines}
}
\description{
Draw conditional effect plot
}
\examples{
fit=lm(justify~skeptic*frame,data=disaster)
condPlot(fit,rangemode=2,xpos=0.7)
condPlot(fit,mode=2,xpos=0.6)
condPlot(fit,mode=3,rangemode=2,xpos=0.5)
fit=lm(mpg~hp*vs,data=mtcars)
condPlot(fit,rangemode=2,xpos=0.6)
condPlot(fit,mode=2,xpos=0.5)
condPlot(fit,mode=3,rangemode=2)
fit=lm(govact~negemot*age+posemot+ideology+sex,data=glbwarm)
condPlot(fit,hjust=c(-0.1,-0.1,1.1))
condPlot(fit,modx.values=c(30,70),hjust=c(-0.1,-0.1,1.1))
condPlot(fit,mode=2,modx.values=c(30,50,70),xpos=0.2)
condPlot(fit,mode=2,xpos=0.2)
condPlot(fit,mode=3,xpos=0.5,hjust=c(-0.1,1.1))
condPlot(fit,pred.values=c(2,3,4),mode=3,xpos=0.5)
}
