% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlot2.R
\name{condPlot2}
\alias{condPlot2}
\title{Draw conditional plot for moderated moderation}
\usage{
condPlot2(fit, mod2 = NULL, mod2.values = NULL, rangemode = 1,
  digits = 3, addlabel = TRUE)
}
\arguments{
\item{fit}{An object of class lm}

\item{mod2}{name of moderator variable}

\item{mod2.values}{values od moderator variable}

\item{rangemode}{integer. 1 or 2}

\item{digits}{integer indicating the number of decimal places}

\item{addlabel}{logical}
}
\description{
Draw conditional plot for moderated moderation
}
\examples{
fit=lm(govact~negemot*sex*age+posemot+ideology,data=glbwarm)
condPlot2(fit)
condPlot2(fit,mod2.values = c(30,50,70))
}
