% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{corPlot}
\alias{corPlot}
\title{Draw correlation plot}
\usage{
corPlot(
  fit,
  label = 2,
  yreverse = TRUE,
  xangle = 45,
  seek = NULL,
  replace = NULL,
  ...
)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{label}{if 0, no label(default), if 1, use r value as label, if 2, use r value with significant mark as label}

\item{yreverse}{Logical. if true, reverse the order of y axis.}

\item{xangle}{axis.x.text.angle}

\item{seek}{string to look for}

\item{replace}{A string of replacement}

\item{...}{Further arguments to be passed on to geom_text}
}
\value{
A ggplot
}
\description{
Draw correlation plot
}
