% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{corTable}
\alias{corTable}
\title{Make a table with correlation}
\usage{
corTable(fit)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}
}
\description{
Make a table with correlation
}
