% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawConcept.R
\name{covar2df}
\alias{covar2df}
\title{Make data.frame with covariates}
\usage{
covar2df(covar = list(), df)
}
\arguments{
\item{covar}{A list}

\item{df}{A data.frame}
}
\description{
Make data.frame with covariates
}
