% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{disaster}
\alias{disaster}
\title{Disaster dataset}
\format{A data.frame with 211 obs. of  5 variables
\describe{
   \item{id}{id}
   \item{frame}{Experimental condition. 0 = naturally caused disaster, 1 = climate change caused disaster}
   \item{donate}{Positive attitudes toward donating}
   \item{justify}{Negative justifications}
   \item{skeptic}{Climate change skepticism}
}}
\source{
Chapman, D. A., & Little, B. (2016). Climate change and disasters: How framing affects justifications for giving or withholding aid to disaster victims. Social Psychological and Personality Science, 7, 13-20.

\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
disaster
}
\description{
Disaster dataset
}
\keyword{datasets}
