% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityTable.R
\name{discriminantValidityTable}
\alias{discriminantValidityTable}
\title{make discriminant Validity Table}
\usage{
discriminantValidityTable(fit)
}
\arguments{
\item{fit}{An object of a class lavaan}
}
\description{
make discriminant Validity Table
}
