% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityTable.R
\name{discriminantValidityTable2}
\alias{discriminantValidityTable2}
\title{make discriminant Validity Table in flextable format}
\usage{
discriminantValidityTable2(fit, vanilla = FALSE)
}
\arguments{
\item{fit}{An object of a class lavaan}

\item{vanilla}{Logical}
}
\description{
make discriminant Validity Table in flextable format
}
