% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMediation.R
\name{divideEquation}
\alias{divideEquation}
\title{divide equation}
\usage{
divideEquation(equation)
}
\arguments{
\item{equation}{a string}
}
\description{
divide equation
}
\examples{
equation="(a1+a3*W)*(b)"
divideEquation(equation)
}
