% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{drawArrows}
\alias{drawArrows}
\title{Draw arrows}
\usage{
drawArrows(
  arrows,
  nodes,
  xmargin = 0.01,
  radx = 0.1,
  rady = 0.04,
  addprime = TRUE
)
}
\arguments{
\item{arrows}{A data.frame}

\item{nodes}{A data.frame}

\item{xmargin}{horizontal margin of plot}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{addprime}{logical Whether add prime to label "c"}
}
\description{
Draw arrows
}
