% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCatModel.R
\name{drawCatModel}
\alias{drawCatModel}
\title{Draw statistical diagram with list of lm result}
\usage{
drawCatModel(fit, labels = NULL, whatLabel = "est", xmargin = 0.01,
  radx = 0.12, ymargin = 0.02, rady = 0.04, maxypos = 0.6,
  minypos = 0, digits = 3)
}
\arguments{
\item{fit}{list of lm object}

\item{labels}{list of variable names}

\item{whatLabel}{What should the edge labels indicate in the path diagram? Choices are c("est","name")}

\item{xmargin}{horizontal margin between nodes}

\item{radx}{horizontal radius of the box.}

\item{ymargin}{vertical margin between nodes}

\item{rady}{vertical radius of the box.}

\item{maxypos}{maximal y position of X or W variables}

\item{minypos}{minimal y position of X or W variables}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Draw statistical diagram with list of lm result
}
\examples{
labels=list(X="protest",W="sexism",M="respappr",Y="liking")
moderator=list(name="sexism",site=list(c("a","c")))
data1=addCatVars(protest,"protest",mode=3)
eq=catMediation(X="protest",M="respappr",Y="liking",moderator=moderator,data=protest,maxylev=6,mode=1)
fit=eq2fit(eq,data=data1)
modelsSummary2(fit,labels=labels)
drawCatModel(fit,labels=labels,whatLabel="name")
drawCatModel(fit,labels=labels)
labels=list(X="protest",W="sexism",M="respappr",Y="liking")
fit=makeCatModel(labels=labels,data=protest)
drawCatModel(fit,labels=labels,maxypos=0.6,minypos=0.2)
drawCatModel(fit,labels=labels,whatLabel="name",maxypos=0.6,minypos=0.2)
}
