% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{drawCovar}
\alias{drawCovar}
\title{Draw covariate}
\usage{
drawCovar(covar = list(), x, y, m, radx = 0.1, rady = 0.06, yinterval = 0.02)
}
\arguments{
\item{covar}{A list}

\item{x}{position of x}

\item{y}{position of y}

\item{m}{position of m}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{yinterval}{vertical interval between box}
}
\description{
Draw covariate
}
