% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModel.R
\name{drawModel}
\alias{drawModel}
\title{Draw statistical diagram with an object of class lavaan or a list of class lm}
\usage{
drawModel(semfit = NULL, labels = NULL, moderator = list(),
  nodelabels = NULL, whatLabel = "name", mode = 1, nodemode = 1,
  xmargin = 0.02, radx = NULL, ymargin = 0.02, xlim = NULL,
  ylim = NULL, box.col = "white", rady = 0.06, maxypos = NULL,
  minypos = 0, ypos = c(1, 0.5), mpos = c(0.5, 0.9),
  xinterval = NULL, yinterval = NULL, xspace = NULL,
  label.pos = list(), interactionFirst = FALSE, totalOnly = FALSE,
  parallel = FALSE, kmediator = FALSE, serial = FALSE, digits = 3)
}
\arguments{
\item{semfit}{An object of class lavaan or a list of class lm}

\item{labels}{list of variable names}

\item{moderator}{A list}

\item{nodelabels}{list of nodes names}

\item{whatLabel}{What should the edge labels indicate in the path diagram? Choices are c("est","name")}

\item{mode}{integer If 1, models with categorical X}

\item{nodemode}{integer If 1, separate node name and node label}

\item{xmargin}{horizontal margin between nodes}

\item{radx}{horizontal radius of the box.}

\item{ymargin}{vertical margin between nodes}

\item{xlim}{the x limits (min,max) of the plot}

\item{ylim}{the y limits (min,max) of the plot}

\item{box.col}{fill color of the box}

\item{rady}{vertical radius of the box.}

\item{maxypos}{maximal y position of X or W variables}

\item{minypos}{minimal y position of X or W variables}

\item{ypos}{The x and y position of Y node. Default value is c(1,0.5)}

\item{mpos}{The x and y position of M node. Default value is c(0.5,0.9)}

\item{xinterval}{numeric. Horizontal intervals among labels for nodes and nodes}

\item{yinterval}{numeric. Vertical intervals among labels for nodes and nodes}

\item{xspace}{numeric. Horizontal distance bewteen nodes}

\item{label.pos}{Optional list of arrow label position}

\item{interactionFirst}{logical If true, place nodes with interaction first}

\item{totalOnly}{logical If true, draw total effect model only}

\item{parallel}{logical If true, draw parallel multiple mediation model}

\item{kmediator}{logical If true, draw parallel multiple mediation model with k mediator}

\item{serial}{Logical. If TRUE, serial variables are added}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Draw statistical diagram with an object of class lavaan or a list of class lm
}
\examples{
library(lavaan)
labels=list(X="frame",W="skeptic",M="justify",Y="donate")
moderator=list(name="skeptic",site=list(c("a","c")))
model=tripleEquation(labels=labels,moderator=moderator,data=disaster)
semfit=sem(model=model,data=disaster)
drawModel(semfit,labels=labels,interactionFirst=TRUE)
labels=list(X="protest",W="sexism",M="respappr",Y="liking")
moderator=list(name="sexism",site=list(c("a","c")))
data1=addCatVars(protest,"protest",mode=3)
model=catMediation(X="protest",M="respappr",Y="liking",moderator=moderator,data=protest,maxylev=6)
semfit=sem(model,data=data1)
nodelabels=list(D1="Ind.Protest",D2="Col.Protest",W="sexism",M="respappr",Y="liking")
drawModel(semfit,labels=labels,nodelabels=nodelabels,whatLabel="name",
       xlim=c(-0.4,1.3))
drawModel(semfit,labels=labels)
labels=list(X="cyl",M=c("am","wt","hp"),Y="mpg",W="vs")
moderator=list(name=c("vs"),site=list(c("a","b")))
model=multipleMediation(labels=labels,moderator=moderator,data=mtcars)
semfit=sem(model=model,data=mtcars)
drawModel(semfit,labels=labels,maxypos=0.5)
labels=list(X="X",M=c("M1","M2","M3"),Y="Y")
nodelabels=c(X="Intervention\\n(vs.control)",
   M=c("Restrained\\nEating","Emotional\\nEating","Perceived\\nBarriers to\\nExercise"),Y="Weight Loss")
drawModel(labels=labels,nodelabels=nodelabels,whatLabel="none",parallel=TRUE,xlim=c(-0.4,1.4),
yinterval=0.02,ylim=c(-0.3,1.2))
labels=list(X="X",M=c("M1","M2","Mk-1","Mk"),Y="Y")
drawModel(labels=labels,parallel=TRUE,kmediator=TRUE,nodemode=2,label.pos=list(c=0.4),radx=0.08)
labels=list(X="cond",M=c("import","pmi"),Y="reaction")
model=multipleMediation(labels=labels,data=pmi,serial=TRUE)
model=multipleMediation(labels=labels,data=pmi)
semfit=sem(model=model,data=pmi)
drawModel(semfit,labels=labels,whatLabel="est",parallel=TRUE)
labels=list(X="X",M=c("M1","M2"),Y="Y")
drawModel(labels=labels,maxypos=0.5,serial=TRUE,nodemode=4)
labels=list(X="X",M=c("M1","M2","M3"),Y="Y")
drawModel(labels=labels,maxypos=0.5,serial=TRUE,nodemode=4,ylim=c(0.4,1))
}
