% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModel2.R
\name{drawModel2}
\alias{drawModel2}
\title{Draw Concept Diagram}
\usage{
drawModel2(labels, vars = NULL, moderator = NULL, nodemode = 1,
  xpos = c(0, 0.5), mpos = c(0.5, 0.9), ypos = c(1, 0.5),
  minypos = 0, maxypos = 0.6, node.pos = list(), serial = FALSE,
  parallel = FALSE, bmatrix = NULL, radx = 0.06, rady = 0.04,
  box.col = "white", xmargin = 0.01, ymargin = 0.01)
}
\arguments{
\item{labels}{A list}

\item{vars}{A list of triple moderators}

\item{moderator}{A list of modeators}

\item{nodemode}{integer If 1, separate node name and node label}

\item{xpos}{The x and y position of X node. Default value is c(0,0.5)}

\item{mpos}{The x and y position of M node. Default value is c(0.5,0.9)}

\item{ypos}{The x and y position of Y node. Default value is c(1,0.5)}

\item{minypos}{minimal y position of X or W variables}

\item{maxypos}{maximal y position of X or W variables}

\item{node.pos}{A optinal list of node position}

\item{serial}{Logical. If TRUE, serial variables are added}

\item{parallel}{logical If true, draw parallel multiple mediation model}

\item{bmatrix}{integer specifying causal relations among mediators}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{box.col}{fill color of the box}

\item{xmargin}{horizontal margin between nodes}

\item{ymargin}{vertical margin between nodes}
}
\description{
Draw Concept Diagram
}
\examples{
labels=list(X="estress",M="affect",Y="withdraw")
vars=list(name=list(c("tenure","age")),site=list(c("a","b")))
moderator=list(name=c("age","sex"),site=list(c("c"),c("b","c")))
drawModel2(labels=labels)
drawModel2(labels=labels,vars=vars)
drawModel2(labels=labels,moderator=moderator)
drawModel2(labels=labels,vars=vars,moderator=moderator)
labels=list(X="X",M=c("M1","M2","M3"),Y="Y")
drawModel2(labels=labels,serial=TRUE)
drawModel2(labels=labels,parallel=TRUE,bmatrix=c(1,1,0,1,0,0,1,1,1,1))
labels=list(X="baby",M=c("wine","tent","sand"),Y="tile")
bmatrix=c(1,1,0,1,0,0,1,1,1,1)
drawModel2(labels=labels,parallel=TRUE,bmatrix=bmatrix)
moderator=list(name=c("milk","hair"),
  matrix=list(c(1,1,0,1,0,0,0,0,0,0),c(0,0,0,0,0,0,0,1,0,0)))
drawModel2(labels=labels,parallel=TRUE,bmatrix=bmatrix,moderator=moderator)
bmatrix=c(1,1,0,0,1,1,1,1,0,1)
moderator=list(name=c("milk","hair"),
            matrix=list(c(1,0,0,0,1,0,1,0,0,0),c(1,1,0,0,0,0,0,0,0,0)),
            pos=c(1,4))
node.pos=list(X=c(0,0.5),M1=c(0.3,0.9),M2=c(0.3,0.1),M3=c(0.7,0.9),
Y=c(1,0.5),W1=c(0.7,0.1),W2=c(0,0.9))
drawModel2(labels=labels,bmatrix=bmatrix,moderator=moderator,node.pos=node.pos)
labels=list(X="baby",M=c("wine","tent","sand"),Y="tile")
vars=list(name=list(c("milk","hair")),matrix=list(c(1,0,0,0,0,0,1,0,0,0)),pos=2)
bmatrix=c(1,1,0,1,0,0,1,1,1,1)
drawModel2(labels=labels,parallel=TRUE,bmatrix=bmatrix,vars=vars)
labels=list(X="X",M=c("M1","M2"),Y="Y")
vars=list(name=list(c("W","Z")),matrix=list(c(0,0,1,0,0,0)),pos=5)
bmatrix=c(1,1,1,1,1,1)
drawModel2(labels=labels,bmatrix=bmatrix,vars=vars)
}
