% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{drawStatDiagram}
\alias{drawStatDiagram}
\title{draw StatDiagram}
\usage{
drawStatDiagram(no, arrows, nodes, labels, xmargin, radx, rady,
  fit = NULL, addprime = TRUE)
}
\arguments{
\item{no}{process macro model number}

\item{arrows}{A data.frame}

\item{nodes}{A data.frame}

\item{labels}{A list}

\item{xmargin}{horizontal margin of plot}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{fit}{An object of class lavaan. Result of lavaan::sem()}

\item{addprime}{logical Whether add prime to label "c"}
}
\description{
draw StatDiagram
}
