% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{drawtext}
\alias{drawtext}
\title{Draw node}
\usage{
drawtext(..., latent = TRUE)
}
\arguments{
\item{...}{Further argument to be passed to textellipse() or textrect()}

\item{latent}{Logical}
}
\description{
Draw node
}
