% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{education}
\alias{education}
\title{Data Set for education and income}
\format{A data.frame with 850 rows and 4 variables:
\describe{
  \item{age}{student age}
  \item{number}{number of students per class}
  \item{duration}{eduation duration}
  \item{income}{income}
}}
\usage{
education
}
\description{
A dataset contains measures about the teacher's knowledge, empathy and intervention about attention-deficit hyperactivity disorder(ADHD).
}
\keyword{datasets}
