% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regEquation.R
\name{eq2fit}
\alias{eq2fit}
\title{Make a list of objects of class lm}
\usage{
eq2fit(equations, data)
}
\arguments{
\item{equations}{equations for linear regression}

\item{data}{A data.frame}
}
\value{
a list of objects of class lm
}
\description{
Make a list of objects of class lm
}
