% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2table.R
\name{eq2var}
\alias{eq2var}
\title{make data.frame with equation}
\usage{
eq2var(eq, labels = list())
}
\arguments{
\item{eq}{equation}

\item{labels}{A list}
}
\description{
make data.frame with equation
}
\examples{
labels=list(X="frame",M="justify",Y="donate",W="skeptic")
eq2var("donate~justify+frame+skeptic+frame:skeptic")
}
