% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2table.R
\name{equations2var}
\alias{equations2var}
\title{make data.frame with equation}
\usage{
equations2var(eq, labels = list())
}
\arguments{
\item{eq}{equation}

\item{labels}{A list}
}
\description{
make data.frame with equation
}
\examples{
labels=list(X="frame",M="justify",Y="donate",W="skeptic")
moderator=list(name="skeptic",site=list(c("a","c")))
eq=multipleMediation(labels=labels,moderator=moderator,mode=1)
covar=list(name=c("C1","C2","C3"),site=list(c("M","Y"),c("M","Y"),c("M","Y")))
eq=multipleMediation(labels=labels,covar=covar,mode=1)
equations2var(eq,labels=labels)
}
