% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{est2Arrows}
\alias{est2Arrows}
\title{Make arrows from estimatesTable}
\usage{
est2Arrows(res)
}
\arguments{
\item{res}{A data.frame, result of estimatesTable}
}
\description{
Make arrows from estimatesTable
}
