% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{est2Nodes}
\alias{est2Nodes}
\title{Make nodes from estimatesTable}
\usage{
est2Nodes(res, lastxno = 2)
}
\arguments{
\item{res}{A data.frame, result of estimatesTable}

\item{lastxno}{A numeric}
}
\description{
Make nodes from estimatesTable
}
