% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{estimatesTable}
\alias{estimatesTable}
\title{convert parameterEstimates to data.frame}
\usage{
estimatesTable(
  fit,
  latent = TRUE,
  regression = TRUE,
  mediation = FALSE,
  covar = FALSE,
  ci = FALSE,
  standardized = TRUE,
  digits = 2
)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{latent}{whether the latent variables be included in result}

\item{regression}{whether the regressions be included in result}

\item{mediation}{whether the mediation effects be included in result}

\item{covar}{whether the covariances be included in result}

\item{ci}{If TRUE, confidence intervals are added to the output}

\item{standardized}{Logical. If TRUE, standardized estimates are added to the output}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\description{
convert parameterEstimates to data.frame
}
