% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{estimatesTable2}
\alias{estimatesTable2}
\title{convert parameterEstimates to flextable}
\usage{
estimatesTable2(
  fit,
  vanilla = TRUE,
  digits = 3,
  seek = NULL,
  replace = NULL,
  ...
)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{vanilla}{Logical}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{seek}{string to look for}

\item{replace}{A string of replacement}

\item{...}{Further arguments to be passed to estimatesTable()}
}
\description{
convert parameterEstimates to flextable
}
