% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{estress}
\alias{estress}
\title{ESTRESS: Economic stress dataset}
\format{
A data.frame with 262 obs. of  7 variables
\describe{
   \item{tenure}{Company Tenure}
   \item{estress}{Economic stress}
   \item{affect}{Depressed affect}
   \item{withdraw}{Withdrawal intentions}
   \item{sex}{Male (1) or Female (0)}
   \item{age}{age}
   \item{ese}{Entrepreneurial self efficacy}
}
}
\source{
Pollack, J., VanEpps, E. M., & Hayes, A. F. (2012). The moderating role of social ties on entrepreneurs' depressed affect and withdrawal intentions in response to economic stress. Journal of Organizational Behavior, 33, 789-810.

\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
estress
}
\description{
ESTRESS: Economic stress dataset
}
\keyword{datasets}
