% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIMM.R
\name{extractIMM}
\alias{extractIMM}
\title{extract index of moderated mediation from string}
\usage{
extractIMM(string)
}
\arguments{
\item{string}{A string}
}
\description{
extract index of moderated mediation from string
}
\examples{
string="(a1+a3*age.mean)*(b1+b3*age.mean)"
string="(a1+a3*skeptic.mean)*(b1+b2*skeptic.mean+b4*Z.mean)"
string="(a1+a3*age.mean)*(b)"
string="(a1+a3*4.12)*(b)"
string="(a)*(b)"
extractIMM(string)
}
