% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{extractLatentVar}
\alias{extractLatentVar}
\title{Extract Latent Variables Data}
\usage{
extractLatentVar(fit, labels)
}
\arguments{
\item{fit}{An object of class lavaan. Result of lavaan::sem()}

\item{labels}{A list}
}
\description{
Extract Latent Variables Data
}
