% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{extractLatentVarName}
\alias{extractLatentVarName}
\title{Extract Latent Variables Names}
\usage{
extractLatentVarName(fit)
}
\arguments{
\item{fit}{An object of class lavaan. Result of lavaan::sem()}
}
\description{
Extract Latent Variables Names
}
