% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIMM.R
\name{extractModerator}
\alias{extractModerator}
\title{Extract name of moderator from string}
\usage{
extractModerator(string)
}
\arguments{
\item{string}{A string}
}
\description{
Extract name of moderator from string
}
\examples{
string="(a1+a3*age.mean)*(b1+b3*age.mean)"
string="(a1+a3*age.mean)*(b)"
string="(a1+a3*4.12)*(b)"
string="(a)*(b)"
extractNumber(string)
extractModerator(string)
}
