% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{extractX}
\alias{extractX}
\title{Extract group by string}
\usage{
extractX(string, groupby = "X")
}
\arguments{
\item{string}{character vector}

\item{groupby}{name of groupby}
}
\description{
Extract group by string
}
