% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{findName}
\alias{findName}
\title{convert name with list}
\usage{
findName(labels, nodeslabels = list(), name = "MiX", exact = FALSE)
}
\arguments{
\item{labels}{A named list}

\item{nodeslabels}{A named list}

\item{name}{A name to look for}

\item{exact}{A logical}
}
\description{
convert name with list
}
\examples{
labels=list(X="wt",M="am",Y="mpg");name="MiX"
nodeslabels=list(X="weight",M="automatic",Y="milepergallon")
findName(labels=labels,nodeslabels=nodeslabels,name="MiX")
findName(labels=labels,name="MiX")
findName(labels=labels,nodeslabels=nodeslabels,name="X")
}
