% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{findNames}
\alias{findNames}
\title{convert a vector of names with list}
\usage{
findNames(labels, names, exact = FALSE)
}
\arguments{
\item{labels}{A named list}

\item{names}{A character vector to look for}

\item{exact}{A logical}
}
\description{
convert a vector of names with list
}
