% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{fit2alpha}
\alias{fit2alpha}
\title{Make a Cronbach alpha table}
\usage{
fit2alpha(fit, digits = 3)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\description{
Make a Cronbach alpha table
}
