% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{fit2df2}
\alias{fit2df2}
\title{Make a data.frame for conceptDiagram}
\usage{
fit2df2(fit)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}
}
\description{
Make a data.frame for conceptDiagram
}
