% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{fit2table}
\alias{fit2table}
\title{Make estimateTable with a list of lm object}
\usage{
fit2table(fit, labels = labels, digits = 3)
}
\arguments{
\item{fit}{A list of lm object}

\item{labels}{A list}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Make estimateTable with a list of lm object
}
\examples{
labels=list(X="frame",M="justify",Y="donate",W="skeptic")
moderator=list(name="skeptic",site=list(c("a","c")))
eq=tripleEquation(labels=labels,moderator=moderator,data=disaster,mode=1)
fit=eq2fit(eq,data=disaster)
fit2table(fit=fit,labels=labels)
}
