% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{fit2vif}
\alias{fit2vif}
\title{Summarizing correlation, tolerance and variance inflation factors of a model}
\usage{
fit2vif(fit, mode = 1, namemode = 1, digits = 3)
}
\arguments{
\item{fit}{An object of class lm}

\item{mode}{integer. one of 1:2}

\item{namemode}{integer. One of 1:3}

\item{digits}{logical}
}
\description{
Summarizing correlation, tolerance and variance inflation factors of a model
}
\examples{
fit=lm(govact~negemot*age,data=glbwarm)
fit2vif(fit)
}
