% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleMediation.R
\name{get2ndIndirect}
\alias{get2ndIndirect}
\title{get2ndIndirect effect}
\usage{
get2ndIndirect(X = NULL, M = NULL, Y = NULL, labels = list())
}
\arguments{
\item{X}{Names of independent variable}

\item{M}{Names of mediator variable}

\item{Y}{Names of dependent variable}

\item{labels}{A list}
}
\description{
get2ndIndirect effect
}
\examples{
get2ndIndirect(X="X",M=c("M1","M2","M3"))
}
