% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalEffectPlot.R
\name{getAspectRatio}
\alias{getAspectRatio}
\title{Get aspect information og a ggplot}
\usage{
getAspectRatio(p)
}
\arguments{
\item{p}{A ggplot object}
}
\description{
Get aspect information og a ggplot
}
