% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary2.R
\name{getCoef}
\alias{getCoef}
\title{Get coef summary table}
\usage{
getCoef(fit)
}
\arguments{
\item{fit}{An object of class lm}
}
\description{
Get coef summary table
}
\examples{
fit=lm(mpg~hp*wt+am,data=mtcars)
getCoef(fit)
}
