% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlot.R
\name{getEq2p}
\alias{getEq2p}
\title{get slope and intercept with 2 points}
\usage{
getEq2p(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x coordinate of the first point}

\item{y1}{y coordinate of the first point}

\item{x2}{x coordinate of the second point}

\item{y2}{y coordinate of the second point}
}
\description{
get slope and intercept with 2 points
}
