% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCatVars.R
\name{getHelmert}
\alias{getHelmert}
\title{Get Helmert Coding of column j of group with length of unique values (count-1)}
\source{
Andrew F. Hayes.(2018) Introduction to Mediation, Moderation and Conditional Process Analysis(2nd Ed.). New York, NY: The Guilford Press. p584
}
\usage{
getHelmert(x, j, count = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{j}{column no}

\item{count}{length unique values of group minus 1}
}
\value{
A numeric vector
}
\description{
Get Helmert Coding of column j of group with length of unique values (count-1)
}
\examples{
x=c(1:4,4:2,2,3,5)
getHelmert(x,1)
getHelmert(mtcars$cyl,1)
}
