% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{getInfo}
\alias{getInfo}
\title{Get information of a model}
\usage{
getInfo(fit, digits = 3)
}
\arguments{
\item{fit}{object of class lm}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Get information of a model
}
\examples{
fit=lm(mpg~wt*hp,data=mtcars)
getInfo(fit)
}
