% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary.R
\name{getMeanSd}
\alias{getMeanSd}
\title{get mean and sd}
\usage{
getMeanSd(data, X, Y, digits)
}
\arguments{
\item{data}{A data.frame}

\item{X}{Name of independant variable}

\item{Y}{Name of dependant variable}

\item{digits}{Integer indicating the number of decimal places}
}
\description{
get mean and sd
}
