% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict2.R
\name{getNewFormula}
\alias{getNewFormula}
\title{Make new formula}
\usage{
getNewFormula(fit, predictors = NULL)
}
\arguments{
\item{fit}{An object of class lm or glm}

\item{predictors}{Names of variables to exclude}
}
\description{
Make new formula
}
\examples{
fit=lm(mpg~hp*wt*cyl+carb+am,data=mtcars)
getNewFormula(fit,predictors=c("wt","cyl"))
}
