% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCatVars.R
\name{getRatioTable}
\alias{getRatioTable}
\title{Get coding table for dummy variables}
\usage{
getRatioTable(count = 3, mode = 1)
}
\arguments{
\item{count}{number of unique length of categorical variable}

\item{mode}{Numeric. One of 1:4. 1= simple indicator coding, 2= sequential coding, 3= Helmert coding, 4= effect coding}
}
\description{
Get coding table for dummy variables
}
\examples{
getRatioTable(count=3)
getRatioTable(count=4,mode=3)
}
