% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSummary3.R
\name{getRepValues}
\alias{getRepValues}
\title{Get representative values}
\usage{
getRepValues(data, colname, rangemode = 2, maxylev = 6, digits = digits)
}
\arguments{
\item{data}{A data.frame}

\item{colname}{Name of column}

\item{rangemode}{Integer. 1 or 2}

\item{maxylev}{Integer. Maximum unique length of variable to be treated as a categorical variable}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Get representative values
}
