% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary.R
\name{getYhat1}
\alias{getYhat1}
\title{Get Yhat value from simple mediation}
\usage{
getYhat1(data, X, M, Y, digits = 3)
}
\arguments{
\item{data}{A data.frame}

\item{X}{Name of independant variable}

\item{M}{Name of moderator variable}

\item{Y}{Name of dependant variable}

\item{digits}{Integer indicating the number of decimal places}
}
\description{
Get Yhat value from simple mediation
}
