% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ggCor}
\alias{ggCor}
\title{Draw a heatmap of correlation test}
\usage{
ggCor(
  data,
  what = 1,
  label = 0,
  colors = NULL,
  title = TRUE,
  mode = 2,
  digits = 2,
  yreverse = TRUE,
  xangle = 45,
  yangle = 0,
  use.label = FALSE
)
}
\arguments{
\item{data}{A data.frame}

\item{what}{if 1, correlation, if 2, partial correlation, if 3, semi-partial correlation}

\item{label}{if 0, no label(default), if 1, use r value as label, if 2, use r value with significant mark as label}

\item{colors}{colors for low, mid and high correlation values}

\item{title}{if true, add title to the heatmap}

\item{mode}{1 or 2}

\item{digits}{The number of decimal place'}

\item{yreverse}{If true, reverse y axis}

\item{xangle}{x-axis text angle}

\item{yangle}{y-axis text angle}

\item{use.label}{Logical whether or not use label in case of labelled data}
}
\description{
Draw a heatmap of correlation test
}
