% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{glbwarm}
\alias{glbwarm}
\title{Global Warming dataset}
\format{
A data.frame with 815 obs. of  7 variables
\describe{
  \item{govact}{Support for government action}
  \item{posemot}{Positive emotions about climate change}
  \item{negemot}{Negative emotions about climate change}
  \item{ideology}{Political ideology (conservatism), 1 = Very Liberal, 2 = Liberal, 3 = Somewhat Liberal, 4 = Moderate; Middle of the Road, 5 = Somewhat Conservative, 6 = Conservative, 7 = Very Conservative}
  \item{age}{Respondent age at last birthday}
  \item{sex}{female(0) or male(1)}
  \item{partyid}{1 = Democrat, 2 = Independent, 3= Republican}
}
}
\source{
\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
glbwarm
}
\description{
Global Warming dataset
}
\keyword{datasets}
