% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{interactStr}
\alias{interactStr}
\title{make interaction equation}
\usage{
interactStr(x, prefix = "a", skip = FALSE, count = 1, addPrefix = TRUE)
}
\arguments{
\item{x}{character vector}

\item{prefix}{prefix}

\item{skip}{whether or not skip}

\item{count}{Numeric}

\item{addPrefix}{A logical}
}
\description{
make interaction equation
}
\examples{
interactStr(LETTERS[1])
interactStr(LETTERS[1:3])
interactStr(LETTERS[1:3],skip=TRUE)
}
