% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict2.R
\name{is.mynumeric}
\alias{is.mynumeric}
\title{Decide whether a vector can be treated as a numeric variable}
\usage{
is.mynumeric(x, maxylev = 6)
}
\arguments{
\item{x}{A vector}

\item{maxylev}{An integer indicating the maximum number of levels of numeric variable be treated as a categorial variable}
}
\description{
Decide whether a vector can be treated as a numeric variable
}
