% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_support.R
\name{is_panel_data}
\alias{is_panel_data}
\title{Check if data is panel data structure}
\usage{
is_panel_data(data, id = NULL, time = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{id}{Character. Name of individual identifier variable}

\item{time}{Character. Name of time identifier variable}
}
\value{
Logical indicating if data has panel structure
}
\description{
Check if data is panel data structure
}
\examples{
data(Grunfeld, package = "plm")
is_panel_data(Grunfeld, id = "firm", time = "year")
}
