% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlot.R
\name{jnPlot}
\alias{jnPlot}
\title{Draw johnson_neyman plot}
\usage{
jnPlot(fit, pred = NULL, modx = NULL, digits = 3, plot = TRUE, ...)
}
\arguments{
\item{fit}{A regression model}

\item{pred}{name of predictor variable}

\item{modx}{name of moderator variable}

\item{digits}{integer indicating the number of decimal places}

\item{plot}{logical. Whether or not draw plot}

\item{...}{Further argumant to be passed to interactions::johnson_neyman()}
}
\description{
Draw johnson_neyman plot
}
\examples{
fit=lm(mpg~hp*wt,data=mtcars)
jnPlot(fit)
fit=lm(justify~skeptic*frame,data=disaster)
res=jnPlot(fit,plot=FALSE)
res$plot
}
