% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModel.R
\name{label2name}
\alias{label2name}
\title{Change label into name}
\usage{
label2name(label, labels)
}
\arguments{
\item{label}{A string}

\item{labels}{A named list}
}
\description{
Change label into name
}
\examples{
label="X:W:Z"
labels=list(X="dep",W="mod",Z="mod2")
label2name(label,labels)
}
