% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2table.R
\name{labels2table}
\alias{labels2table}
\title{Make table with labels}
\usage{
labels2table(labels = labels, moderator = list(), serial = FALSE)
}
\arguments{
\item{labels}{A list}

\item{moderator}{A list}

\item{serial}{A logical}
}
\description{
Make table with labels
}
\examples{
labels=list(X="frame",M="justify",Y="donate",W="skeptic")
moderator=list(name="skeptic",site=list(c("a","c")))
labels=list(X="X",M=c("M1","M2","M3"),Y="Y")
labels2table(labels,serial=TRUE)
labels2table(labels,moderator=moderator)
}
