% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlotCat.R
\name{makeAnovaDf}
\alias{makeAnovaDf}
\title{Make data summarizing ANOVA results}
\usage{
makeAnovaDf(labels, data, maxylev = 6)
}
\arguments{
\item{labels}{Named list of variables}

\item{data}{A data.frame}

\item{maxylev}{maximal unique length of categorical variable}
}
\description{
Make data summarizing ANOVA results
}
\examples{
labels=list(X="protest",W="sexism",Y="liking")
makeAnovaDf(labels=labels,data=protest)
}
