% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCatEquation.R
\name{makeCatEquation3}
\alias{makeCatEquation3}
\title{Make equation for sem and lm for multiple X or multiple Y}
\usage{
makeCatEquation3(X = NULL, Y = NULL, W = NULL, labels = list(),
  prefix = "b", mode = 0, pos = list(), bmatrix, depy = FALSE,
  depx = FALSE)
}
\arguments{
\item{X}{Names of independent variable}

\item{Y}{Names of dependent variable}

\item{W}{Names of moderators}

\item{labels}{optional list}

\item{prefix}{a character}

\item{mode}{A numeric}

\item{pos}{Numeric moderator position}

\item{bmatrix}{integer specifying causal relations among mediators}

\item{depy}{logical}

\item{depx}{logical}
}
\description{
Make equation for sem and lm for multiple X or multiple Y
}
\examples{
cat(makeCatEquation3(X="X",Y=c("M1","M2","M3"),prefix="a",bmatrix=c(1,1,0,1,0,0,1,1,1,1)))
cat(makeCatEquation3(X="X",Y=c("M1","M2","M3"),prefix="a",bmatrix=c(1,1,0,1,0,1,1,1,1,1)))
cat(makeCatEquation3(X="X",Y=c("M1","M2","M3"),prefix="a",bmatrix=c(1,1,0,1,1,0,1,1,1,1)))
cat(makeCatEquation3(X="X",Y=c("M1","M2","M3"),prefix="a",bmatrix=c(1,1,1,1,1,1,1,1,1,1)))
cat(makeCatEquation3(X=c("M1","M2","M3"),Y="Y",prefix="a",bmatrix=c(1,1,1,1,1,1,1,1,1,1),depy=TRUE))
cat(makeCatEquation3(X="X",Y="Y",prefix="a",bmatrix=c(1,1,1,1,1,1,1,1,1,1),depy=TRUE,depx=TRUE))
cat(makeCatEquation3(X="X",Y="Y",prefix="a",bmatrix=c(1,1,1,1,1,1,0,1,1,1),depy=TRUE,depx=TRUE))
cat(makeCatEquation3(X=c("M1","M2"),Y="Y",prefix="a",bmatrix=c(1,1,1,0,1,0),depy=TRUE))
}
