% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlotCat.R
\name{makeCEDf}
\alias{makeCEDf}
\title{Make data summarizing conditional effects}
\usage{
makeCEDf(labels = labels, data, maxylev = 6)
}
\arguments{
\item{labels}{Named list of variables}

\item{data}{A data.frame}

\item{maxylev}{maximal unique length of categorical variable}
}
\description{
Make data summarizing conditional effects
}
\examples{
labels=list(X="protest",W="sexism",Y="liking")
makeCEDf(labels=labels,data=protest)
}
