% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{makeCoefLabel}
\alias{makeCoefLabel}
\title{Change regression coefficient name}
\usage{
makeCoefLabel(name, dep, labels, constant, prefix)
}
\arguments{
\item{name}{string vector to change}

\item{dep}{names of dependent variable}

\item{labels}{optional list}

\item{constant}{name of constant}

\item{prefix}{name of prefix}
}
\description{
Change regression coefficient name
}
