% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{makeEquation}
\alias{makeEquation}
\title{Make mediation equations 3}
\usage{
makeEquation(X, M, Y, add2ndMediation = TRUE, covar = list())
}
\arguments{
\item{X}{A character vectors indicating independent variables}

\item{M}{A character vectors indicating mediators}

\item{Y}{A character vectors indicating dependent variables}

\item{add2ndMediation}{whether or not make a 2nd mediation equation}

\item{covar}{Optional list of covariates}
}
\description{
Make mediation equations 3
}
\examples{
X="X";M=c("M1","M2","M3");Y=c("Y1","Y2");add2ndMediation=TRUE
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M1","Y1"),"Y2","Y2"))
cat(makeEquation(X,M,Y,covar=covar))
}
