% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{makeEquation1}
\alias{makeEquation1}
\title{Make mediation equations 1}
\usage{
makeEquation1(X, M, stage = 1, start = 0, add2ndMediation = TRUE)
}
\arguments{
\item{X}{A character vectors indicating independent variables}

\item{M}{A character vectors indicating mediators}

\item{stage}{An integer indicating the order}

\item{start}{An integer}

\item{add2ndMediation}{whether or not make a 2nd mediation equation}
}
\description{
Make mediation equations 1
}
