% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{makeEquation2}
\alias{makeEquation2}
\title{Make mediation equations 2}
\usage{
makeEquation2(X, M, Y)
}
\arguments{
\item{X}{A character vectors indicating independent variables}

\item{M}{A character vectors indicating mediators}

\item{Y}{A character vectors indicating dependent variables}
}
\description{
Make mediation equations 2
}
