% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{makeEquation3}
\alias{makeEquation3}
\title{Make mediation equations 3}
\usage{
makeEquation3(X, M, Y, add2ndMediation = TRUE)
}
\arguments{
\item{X}{A character vectors indicating independent variables}

\item{M}{A character vectors indicating mediators}

\item{Y}{A character vectors indicating dependent variables}

\item{add2ndMediation}{whether or not make a 2nd mediation equation}
}
\description{
Make mediation equations 3
}
