% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMediation.R
\name{makeIndirectEquationCat}
\alias{makeIndirectEquationCat}
\title{Make indirect equation for categorical variables}
\usage{
makeIndirectEquationCat(X, M, temp1, temp2, temp3, moderatorNames,
  range = TRUE, data = NULL, rangemode = 1, probs = c(0.16, 0.5,
  0.84))
}
\arguments{
\item{X}{A character string}

\item{M}{A character string}

\item{temp1}{A character vector}

\item{temp2}{A character vector}

\item{temp3}{A character vector}

\item{moderatorNames}{A character vector}

\item{range}{A logical}

\item{data}{A data.frame}

\item{rangemode}{range mode}

\item{probs}{numeric vector of probabilities with values in [0,1]}
}
\description{
Make indirect equation for categorical variables
}
\examples{
X= c("d1","d2");M="am"
temp1=c("a1*d1","a2*d2","a3*wt","a4*d1:wt","a5*d2:wt")
temp2="b1*am"
temp3=c("c1*d1","c2*d2","c3*wt","c4*d1:wt","c5*d2:wt")
moderatorNames="wt"
cat(makeIndirectEquationCat(X,M,temp1,temp2,temp3,moderatorNames,range=TRUE))
}
