% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlot2.R
\name{makeLabel}
\alias{makeLabel}
\title{Make Labels}
\usage{
makeLabel(fit, mod2.values = c(30, 50, 70), digits = 3)
}
\arguments{
\item{fit}{An object of class lm}

\item{mod2.values}{values od moderator variable}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Make Labels
}
