% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTx.R
\name{makePPTx}
\alias{makePPTx}
\title{make powerpoint presentation}
\usage{
makePPTx(
  data,
  preprocessing = "",
  filename = "report.pptx",
  rawDataName = NULL,
  rawDataFile = "rawData.RDS",
  rmdRemove = TRUE
)
}
\arguments{
\item{data}{A data.frame with title and code}

\item{preprocessing}{string preprocessing}

\item{filename}{character name of output file}

\item{rawDataName}{The name of the rawData}

\item{rawDataFile}{The name of the rawData file which the data are to be read from}

\item{rmdRemove}{A logical}
}
\description{
make powerpoint presentation
}
