% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{matrix2df}
\alias{matrix2df}
\title{Make data.frame with matrix}
\usage{
matrix2df(matrix = c(1, 1, 1, 0, 0, 1), labels = NULL)
}
\arguments{
\item{matrix}{a numeric vector}

\item{labels}{Optional list of labels}
}
\description{
Make data.frame with matrix
}
\examples{
labels=list(X="indep",M=c("med1","med2"),Y="dep")
matrix2df(c(1,1,1,0,0,1),labels=labels)
}
