% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{matrix2no}
\alias{matrix2no}
\title{Calculate the dimension of matrix}
\usage{
matrix2no(matrix = c(1, 1, 1, 0, 0, 1))
}
\arguments{
\item{matrix}{a numeric vector}
}
\description{
Calculate the dimension of matrix
}
\examples{
matrix2no(c(1,1,1,0,0,1))
}
