% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{meanCentering}
\alias{meanCentering}
\title{Perform mean centering}
\usage{
meanCentering(data, names)
}
\arguments{
\item{data}{A data.frame}

\item{names}{column names to mean centering}
}
\description{
Perform mean centering
}
\examples{
library(processR)
newData=meanCentering(education,colnames(education)[1:3])
}
