% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary.R
\name{meanSummaryTable}
\alias{meanSummaryTable}
\title{Make mean summary table}
\usage{
meanSummaryTable(..., vanilla = TRUE)
}
\arguments{
\item{...}{Further arguments to be passed to meanSummary}

\item{vanilla}{logical}
}
\description{
Make mean summary table
}
\examples{
labels=list(X="cond",Y="reaction",M="pmi")
xlabels=c("Front Page","Interior Page")
meanSummaryTable(data=pmi,labels=labels,xlabels=xlabels)
labels=list(X="frame",Y="justify",W="skeptic")
xlabels=c("Natural causes condition","Climate change condition")
meanSummaryTable(data=disaster,labels=labels,xlabels=xlabels)
}
