% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{medSummaryTable2}
\alias{medSummaryTable2}
\title{Make a table summarizing the mediation effects}
\usage{
medSummaryTable2(x, vanilla = TRUE)
}
\arguments{
\item{x}{An object of class medSummary2}

\item{vanilla}{A logical}
}
\description{
Make a table summarizing the mediation effects
}
