% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityTable.R
\name{modelFitGuideTable2}
\alias{modelFitGuideTable2}
\title{Model fit guide table}
\usage{
modelFitGuideTable2(vanilla = FALSE)
}
\arguments{
\item{vanilla}{Logical}
}
\description{
Model fit guide table
}
