% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{modelFitTable}
\alias{modelFitTable}
\title{Extract model fit measures to data.frame}
\usage{
modelFitTable(fit, digits = 2, names = NULL)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{names}{names of statistic to be extracted}
}
\value{
A data.frame
}
\description{
Extract model fit measures to data.frame
}
