% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{modelFitTable2}
\alias{modelFitTable2}
\title{Extract model fit measures to flextable}
\usage{
modelFitTable2(fit, vanilla = FALSE, ...)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{vanilla}{Logical}

\item{...}{Further arguments to be passed to modelFitTable()}
}
\description{
Extract model fit measures to flextable
}
