% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{modelsSummary}
\alias{modelsSummary}
\title{Make Model Coef Summary}
\usage{
modelsSummary(fit, labels = NULL)
}
\arguments{
\item{fit}{A list of objects of class lm}

\item{labels}{optional list}
}
\value{
A data.frame
}
\description{
Make Model Coef Summary
}
\examples{
fit1=lm(mpg~wt,data=mtcars)
fit2=lm(mpg~wt*hp,data=mtcars)
fit3=lm(mpg~wt*hp*am,data=mtcars)
labels=list(Y="mpg",X="wt",W="hp",Z="am")
modelsSummary(list(fit1),labels=labels)
modelsSummary(list(fit1,fit2),labels=labels)
modelsSummary(list(fit1,fit2))
modelsSummary(list(fit1,fit2,fit3),labels=labels)
}
