% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary2.R
\name{modelsSummary2}
\alias{modelsSummary2}
\title{Make Summary for Model Coefficients}
\usage{
modelsSummary2(fit, labels = NULL, prefix = "b", fitlabels = NULL)
}
\arguments{
\item{fit}{A list of objects of class lm}

\item{labels}{optional list}

\item{prefix}{A character}

\item{fitlabels}{Optional. labels of models}
}
\value{
A data.frame
}
\description{
Make Summary for Model Coefficients
}
\examples{
fit1=lm(mpg~wt,data=mtcars)
fit2=lm(mpg~wt*hp,data=mtcars)
labels=list(Y="mpg",X="wt",W="hp",Z="am")
fit=list(fit1,fit2)
modelsSummary2(fit,labels=labels)
}
