% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary2.R
\name{modelsSummary2Table}
\alias{modelsSummary2Table}
\title{Make Summary Table 2 for Model Coefficients}
\usage{
modelsSummary2Table(x, vanilla = TRUE, ...)
}
\arguments{
\item{x}{An object of class modelSummary2}

\item{vanilla}{A logical}

\item{...}{further arguments to be passed to modelsSummary2()}
}
\description{
Make Summary Table 2 for Model Coefficients
}
