% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawConcept.R
\name{moderator2df}
\alias{moderator2df}
\title{Make data.frame from a list of moderator}
\usage{
moderator2df(moderator, mpos = c(0.5, 0.9), vars = NULL, df = NULL)
}
\arguments{
\item{moderator}{A list}

\item{mpos}{A numeric vector of length 2}

\item{vars}{A list}

\item{df}{A data.frame}
}
\description{
Make data.frame from a list of moderator
}
\examples{
moderator=list(name=c("milk","hair"),matrix=list(c(1,1,0,1,0,0,0,0,0,0)
   ,c(0,0,0,0,0,0,0,1,0,0)))
moderator2df(moderator)
}
