% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{moderator2pos}
\alias{moderator2pos}
\title{get position from moderator}
\usage{
moderator2pos(moderator = list(), x, y, m)
}
\arguments{
\item{moderator}{A list}

\item{x}{position of x}

\item{y}{position of y}

\item{m}{position of m}
}
\description{
get position from moderator
}
