% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{modmedSummary2Table}
\alias{modmedSummary2Table}
\title{make table summarizing moderated mediation effect}
\usage{
modmedSummary2Table(x, vanilla = TRUE, ...)
}
\arguments{
\item{x}{An object of class lavaan or modmedSummary2}

\item{vanilla}{logical.}

\item{...}{Further arguments to be passed to modmedSummary}
}
\description{
make table summarizing moderated mediation effect
}
