% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{modmedSummaryTable}
\alias{modmedSummaryTable}
\title{Make a table summarizing the moderated mediation}
\usage{
modmedSummaryTable(x, vanilla = TRUE)
}
\arguments{
\item{x}{An object of class modmedSummary}

\item{vanilla}{A logical}
}
\description{
Make a table summarizing the moderated mediation
}
