% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSummary.R
\name{modSummary}
\alias{modSummary}
\title{Make table summarizing moderation effect}
\usage{
modSummary(fit, pred = NULL, modx = NULL, pred.values = NULL,
  modx.values = NULL, rangemode = 2, digits = 3)
}
\arguments{
\item{fit}{An object of class lm}

\item{pred}{name of predictor variable}

\item{modx}{name of moderator variable}

\item{pred.values}{Values of predictor variables}

\item{modx.values}{Values of modifier variables}

\item{rangemode}{integer. 1 or 2}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Make table summarizing moderation effect
}
\examples{
fit=lm(justify~frame*skeptic,data=disaster)
modSummary(fit)
}
