% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSummary3.R
\name{modSummary3Table}
\alias{modSummary3Table}
\title{Make flextable summarizing moderation effect}
\usage{
modSummary3Table(x, vanilla = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{vanilla}{logical}

\item{...}{Further argument to be passed to modSummary3}
}
\description{
Make flextable summarizing moderation effect
}
\examples{
fit=lm(govact~negemot*sex+negemot*age+posemot+ideology,data=glbwarm)
modSummary3Table(fit,mod2.values=c(30,50,70))
}
