% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSummary.R
\name{modSummaryTable}
\alias{modSummaryTable}
\title{Make flextable summarizing moderation effect}
\usage{
modSummaryTable(x, vanilla = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{vanilla}{logical}

\item{...}{Further argument to be passed to modSummary}
}
\description{
Make flextable summarizing moderation effect
}
\examples{
fit=lm(justify~skeptic*frame,data=disaster)
modSummaryTable(fit)
}
