% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{moreModels}
\alias{moreModels}
\title{more models data}
\format{A data.frame 2 variables
\describe{
  \item{no}{process macro model number}
  \item{no1}{model number}
}}
\usage{
moreModels
}
\description{
more models data
}
\keyword{datasets}
