% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleMediation.R
\name{multipleMediation}
\alias{multipleMediation}
\title{Make Mediation Equation with multiple X or multiple Y}
\usage{
multipleMediation(X = NULL, M = NULL, Y = NULL, labels = list(),
  data, moderator = list(), covar = NULL, mode = 0, range = TRUE,
  rangemode = 1)
}
\arguments{
\item{X}{Names of independent variable}

\item{M}{Names of mediator variable}

\item{Y}{Names of dependent variable}

\item{labels}{optional list}

\item{data}{A data.frame}

\item{moderator}{A list}

\item{covar}{A list of covariates}

\item{mode}{A numeric. 0: SEM equation, 1: regression equation}

\item{range}{A logical}

\item{rangemode}{range mode}
}
\description{
Make Mediation Equation with multiple X or multiple Y
}
\examples{
labels=list(X=c("cyl","wt"),M="am",Y="mpg")
moderator=list(name=c("vs"),site=list(c("a1","b1")))
covar=list(name=c("carb","disp"),label=c("carb","disp"),site=list(c("M","Y"),"Y","Y"))
cat(multipleMediation(labels=labels,data=mtcars))
cat(multipleMediation(labels=labels,moderator=moderator,data=mtcars))
labels=list(X="wt",M=c("cyl","am"),Y="mpg")
moderator=list(name=c("vs"),site=list(c("a1","b1")))
cat(multipleMediation(labels=labels,data=mtcars,range=FALSE))
cat(multipleMediation(labels=labels,moderator=moderator,data=mtcars,range=FALSE))
labels=list(X="X",M=c("M1","M2"),Y="Y")
moderator=list(name=c("W"),site=list(c("a2","b1")))
moderator=list(name=c("W"),site=list(c("a","b")))
cat(multipleMediation(labels=labels,moderator=moderator,range=FALSE))
cat(multipleMediation(labels=labels,moderator=moderator,data=mtcars,range=FALSE))
cat(multipleMediation(X="am",Y="mpg",data=mtcars,moderator=moderator,covar=covar))
}
