% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{myarrow2}
\alias{myarrow2}
\title{Draw arrow with adjustment of a position}
\usage{
myarrow2(
  nodes,
  from,
  to,
  label = "",
  no,
  radx = 0.12,
  rady = 0.04,
  xmargin = 0.01,
  label.pos = 0.5,
  arr.pos = NULL,
  addprime = TRUE,
  xspace = NULL,
  mode = 1,
  ...
)
}
\arguments{
\item{nodes}{A data.frame}

\item{from}{coordinates (x,y) of the point *from* which to draw arrow.}

\item{to}{coordinates (x,y) of the point *to* which to draw arrow.}

\item{label}{label to display}

\item{no}{process macro model number}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{xmargin}{horizontal margin of plot}

\item{label.pos}{label position}

\item{arr.pos}{arrow position}

\item{addprime}{logical Whether add prime to label "c"}

\item{xspace}{numeric horizontal space between nodes}

\item{mode}{integer mode for adjustxpos}

\item{...}{Further argument to be passed to straightarrow()}
}
\description{
Draw arrow with adjustment of a position
}
