% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTx.R
\name{mycat}
\alias{mycat}
\title{append something to file}
\usage{
mycat(..., file = "report.Rmd")
}
\arguments{
\item{...}{Further argument passed to the cat()}

\item{file}{name of file}
}
\description{
append something to file
}
