% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mycor}
\alias{mycor}
\title{Perform correlation and linear regression for a data.frame}
\usage{
mycor(x, digits = 3)
}
\arguments{
\item{x}{A data.frame}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Perform correlation and linear regression for a data.frame
}
