% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{myformat}
\alias{myformat}
\title{Format a numeric vector}
\usage{
myformat(x, digits = 3)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Format a numeric vector
}
