% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mylm}
\alias{mylm}
\title{Correlation and Fitting linear model function for function "mycor"}
\usage{
mylm(y, x, digits = 3)
}
\arguments{
\item{y}{numeric vectors of data values}

\item{x}{numeric vectors of data values}

\item{digits}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used.}
}
\value{
mylm returns a list of following components

    \describe{
       \item{out}{a list of class "htest" from \code{\link{cor.test}}
          between the last paired samples in a data.frame.}
       \item{result}{a numeric vector of length 4, consist of r and p values
             from \code{\link{cor.test}},slope and intercept values from
             \code{\link{lm}} between numeric vector y and x}
    }
}
\description{
Correlation and Fitting linear model function for function "mycor"
}
