% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Node Data Set for drawing statistical diagram of process macro model}
\format{
A data.frame with 327 rows and 4 variables
\describe{
  \item{no}{process macro model number}
  \item{name}{name of node}
  \item{xpos}{x position}
  \item{ypos}{y position}
}
}
\usage{
nodes
}
\description{
Node Data Set for drawing statistical diagram of process macro model
}
\keyword{datasets}
