% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_support.R
\name{panel_multipleMediation}
\alias{panel_multipleMediation}
\title{Panel data multiple mediation analysis}
\usage{
panel_multipleMediation(
  X,
  Y,
  M,
  id,
  time,
  data,
  moderator = list(),
  model = "within",
  effect = "individual",
  robust = TRUE,
  lag = 0,
  ...
)
}
\arguments{
\item{X}{Names of independent variable(s)}

\item{Y}{Name of dependent variable}

\item{M}{Names of mediator variable(s)}

\item{id}{Character. Name of individual identifier variable}

\item{time}{Character. Name of time identifier variable}

\item{data}{A data.frame}

\item{moderator}{A list specifying moderator variables}

\item{model}{Character. Panel model type: "within", "random", "pooling", "between"}

\item{effect}{Character. Effects to include: "individual", "time", "twoways"}

\item{robust}{Logical. Whether to use robust standard errors}

\item{lag}{Integer. Number of lags for dynamic panel models}

\item{...}{Additional arguments passed to plm functions}
}
\value{
A list containing panel mediation analysis results
}
\description{
Panel data multiple mediation analysis
}
\examples{
\dontrun{
data(Grunfeld, package = "plm")
# Add some example variables for demonstration
Grunfeld$mediator <- rnorm(nrow(Grunfeld))
result <- panel_multipleMediation(
  X = "inv", Y = "value", M = "mediator",
  id = "firm", time = "year", data = Grunfeld,
  model = "within"
)
}
}
