% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleMediation.R
\name{parallelMatrix}
\alias{parallelMatrix}
\title{Make bmatrix for parallel multiple mediator model}
\usage{
parallelMatrix(n = 2)
}
\arguments{
\item{n}{integer number of mediator}
}
\description{
Make bmatrix for parallel multiple mediator model
}
\examples{
parallelMatrix(3)
}
