% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{parrows}
\alias{parrows}
\title{Arrow Data Set for drawing statistical diagram of process macro model}
\format{
A data.frame with 392 rows and 6 variables
\describe{
  \item{no}{process macro model number}
  \item{name}{name of arrow}
  \item{start}{start node}
  \item{end}{end node}
  \item{labelpos}{position of label}
  \item{arrowpos}{position of arrow head}
}
}
\usage{
parrows
}
\description{
Arrow Data Set for drawing statistical diagram of process macro model
}
\keyword{datasets}
