% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{pastecolon}
\alias{pastecolon}
\title{paste two character with colon}
\usage{
pastecolon(temp, x)
}
\arguments{
\item{temp}{a character}

\item{x}{a character}
}
\description{
paste two character with colon
}
