% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationBK.R
\name{plot.mediationBK}
\alias{plot.mediationBK}
\title{S3 method for class mediationBK}
\usage{
\method{plot}{mediationBK}(x, ...)
}
\arguments{
\item{x}{An object of class mediationBK}

\item{...}{Further arguments to be passed to plot()}
}
\description{
S3 method for class mediationBK
}
