% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlot.R
\name{plotCoef}
\alias{plotCoef}
\title{Make Slopes Plot}
\usage{
plotCoef(ss, color = "deepskyblue2", size = 0.75, digits = 1)
}
\arguments{
\item{ss}{An object of class sim_slopes}

\item{color}{Name of color}

\item{size}{size of pointrange}

\item{digits}{An integer indicating the number of decimal places}
}
\value{
A ggplot
}
\description{
Make Slopes Plot
}
