% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{pmacro}
\alias{pmacro}
\title{Data Set for process macro model}
\format{
A data.frame with 43 rows and 7 variables
\describe{
  \item{no}{process macro model number}
  \item{X}{name of independent variable}
  \item{M}{names of mediator variables}
  \item{Y}{name of dependent variable}
  \item{modName}{names of moderator variables}
  \item{modSite}{sites of moderators}
  \item{pos}{position of moderators}
}
}
\usage{
pmacro
}
\description{
Data Set for process macro model
}
\keyword{datasets}
