% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{pmi}
\alias{pmi}
\title{PMI: Presumed Media Influence dataset}
\format{A data.frame with 123 obs. of 6 variables
\describe{
   \item{cond}{front (1) or interior (0) page of the newspaper}
   \item{pmi}{presumed media influence}
   \item{import}{article is on an important topic}
   \item{reaction}{sugar purchase}
   \item{gender}{GENDER: female (0) or male (1)}
   \item{age}{age}
}}
\source{
Tal-Or, N., Cohen, J., Tsafati, Y., & Gunther, A. C. (2010). Testing causal direction in the influence of presumed media influence. Communication Research, 37, 801-824.

\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
pmi
}
\description{
PMI: Presumed Media Influence dataset
}
\keyword{datasets}
