% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{print.compareVIF}
\alias{print.compareVIF}
\title{S3 method of class compareVIF}
\usage{
\method{print}{compareVIF}(x, ...)
}
\arguments{
\item{x}{An object of class compareVIF}

\item{...}{Further arguments to be passed to print}
}
\description{
S3 method of class compareVIF
}
