% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary.R
\name{print.meanSummary}
\alias{print.meanSummary}
\title{S3 method of class meanSummary}
\usage{
\method{print}{meanSummary}(x, ...)
}
\arguments{
\item{x}{An object of class meanSummary}

\item{...}{Further argumants to be passed to print()}
}
\description{
S3 method of class meanSummary
}
