% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationBK.R
\name{print.mediationBK}
\alias{print.mediationBK}
\title{S3 method for class mediationBK}
\usage{
\method{print}{mediationBK}(x, ...)
}
\arguments{
\item{x}{An object of class mediationBK}

\item{...}{Further arguments to be passed to print()}
}
\description{
S3 method for class mediationBK
}
