% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{print.medSummary}
\alias{print.medSummary}
\title{S3 method print for an object of class modmedSummary}
\usage{
\method{print}{medSummary}(x, ...)
}
\arguments{
\item{x}{An object of class medSummary}

\item{...}{additional arguments to pass to print.medSummary}
}
\description{
S3 method print for an object of class modmedSummary
}
