% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{print.modelSummary}
\alias{print.modelSummary}
\title{S3 method print for object modelSummary}
\usage{
\method{print}{modelSummary}(x, ...)
}
\arguments{
\item{x}{Object of class modelSummary}

\item{...}{additional arguments to pass to print.modelSummary}
}
\description{
S3 method print for object modelSummary
}
