% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary2.R
\name{print.modelSummary2}
\alias{print.modelSummary2}
\title{S3 method print for object modelSummary2}
\usage{
\method{print}{modelSummary2}(x, ...)
}
\arguments{
\item{x}{Object of class modelSummary}

\item{...}{additional arguments to pass to print.modelSummary}
}
\description{
S3 method print for object modelSummary2
}
