% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{print.modmedSummary}
\alias{print.modmedSummary}
\title{S3 method print for an object of class modmedSummary}
\usage{
\method{print}{modmedSummary}(x, ...)
}
\arguments{
\item{x}{An object of class modmedSummary}

\item{...}{additional arguments to pass to print.modmedSummary}
}
\description{
S3 method print for an object of class modmedSummary
}
