% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{print.modmedSummary2}
\alias{print.modmedSummary2}
\title{S3 method print for an object of class modmedSummary2}
\usage{
\method{print}{modmedSummary2}(x, ...)
}
\arguments{
\item{x}{An object of class modmedSummary2}

\item{...}{additional arguments to pass to print.modmedSummary2}
}
\description{
S3 method print for an object of class modmedSummary2
}
