% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSummary2.R
\name{print.modSummary2}
\alias{print.modSummary2}
\title{S3 method of print}
\usage{
\method{print}{modSummary2}(x, ...)
}
\arguments{
\item{x}{An object of class modSummary2}

\item{...}{Further argument to be passed to print()}
}
\description{
S3 method of print
}
