% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlotCat2.R
\name{productEq}
\alias{productEq}
\title{Make products of equations}
\usage{
productEq(equation1, equation2)
}
\arguments{
\item{equation1}{The first equation}

\item{equation2}{Thw second equation}
}
\description{
Make products of equations
}
\examples{
equation1=c("a1+b1*W")
equation2=c("a2+b2*W")
productEq(equation1,equation2)
}
