% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{protest}
\alias{protest}
\title{Protest dataset}
\format{A data.frame with 129 rows and 6 variables
\describe{
   \item{subnum}{subject number}
   \item{protest}{experimental condition, 0 = no protest, 1 = individual protest, 2 = group protest}
   \item{sexism}{perceived pervasiveness of sex discrimination}
   \item{angry}{anger toward the attorney}
   \item{liking}{liking of the attorney}
   \item{respappr}{appropriateness of response}
}}
\source{
Garcia, D. M., Schmitt, M. T., Branscombe, N. R., & Ellemers, N. (2010). Women's reactions to ingroup members who protest discriminatory treatment: The importance of beliefs about inequality and response appropriateness. European Journal of Social Psychology, 40, 733-745.

\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
protest
}
\description{
Protest dataset
}
\keyword{datasets}
