% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regEquation.R
\name{regEquation}
\alias{regEquation}
\title{Make regression equation}
\usage{
regEquation(
  X = "X",
  M = NULL,
  Y = "Y",
  moderator = list(),
  covar = list(),
  secondIndirect = FALSE
)
}
\arguments{
\item{X}{A character vectors indicating independent variables}

\item{M}{A character vectors indicating mediators}

\item{Y}{A character vectors indicating dependent variables}

\item{moderator}{moderator}

\item{covar}{covariates}

\item{secondIndirect}{A logical}
}
\description{
Make regression equation
}
\examples{
X="X";M=NULL;Y="Y"; moderator=list(name="W",site=list("c"))
regEquation(X,M,Y,moderator)
M=c("M1","M2")
regEquation(X,M,Y,moderator,secondIndirect=TRUE)
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M1","Y"),"Y","Y"))
regEquation(X,M,Y,moderator,covar=covar)
covar=list(name=c("ese","sex","tenure"),site=list(c("M","Y"),c("M","Y"),c("M","Y")))
regEquation(X="estress",M="affect",Y="withdraw",covar=covar)
}
