% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityTable.R
\name{reliabilityTable2}
\alias{reliabilityTable2}
\title{make reliability Table in flextable format}
\usage{
reliabilityTable2(fit, vanilla = FALSE)
}
\arguments{
\item{fit}{An object of a class lavaan}

\item{vanilla}{Logical}
}
\description{
make reliability Table in flextable format
}
