% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{rightPrint}
\alias{rightPrint}
\title{Print a string in center}
\usage{
rightPrint(string, width)
}
\arguments{
\item{string}{A string}

\item{width}{A numeric}
}
\description{
Print a string in center
}
