% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{seekGroup1}
\alias{seekGroup1}
\title{Find group with variable name}
\usage{
seekGroup1(var, res)
}
\arguments{
\item{var}{A string to seek}

\item{res}{A data.frame. Result of parameterEstimates function of package lavaan or subset.}
}
\description{
Find group with variable name
}
