% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{seekGroup2}
\alias{seekGroup2}
\title{Find group with variable name}
\usage{
seekGroup2(var, res, group)
}
\arguments{
\item{var}{A string to seek}

\item{res}{A data.frame. Result of parameterEstimates function of package lavaan or subset.}

\item{group}{A character vector}
}
\description{
Find group with variable name
}
