% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{seekNameVars}
\alias{seekNameVars}
\title{select names of variables from list var}
\usage{
seekNameVars(vars, site = "a")
}
\arguments{
\item{vars}{A list}

\item{site}{Site for look for}
}
\description{
select names of variables from list var
}
\examples{
vars=list(name=list(c("W","Z"),c("V","Q")),site=list(c("a","c"),c("b","c")))
vars=list(name=list(c("W","Z")),site=list(c("a","c")))
seekNameVars(vars,"a")
seekNameVars(vars,"b")
seekNameVars(vars,"c")
}
