% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{seekVar}
\alias{seekVar}
\title{Seek var form covariates}
\usage{
seekVar(covar = list(), var, prefix = "h", start = 1)
}
\arguments{
\item{covar}{A list of covariates}

\item{var}{A name of variable to look for}

\item{prefix}{A prefix}

\item{start}{A start number}
}
\description{
Seek var form covariates
}
\examples{
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M","Y"),"Y","Y"))
var="M"
seekVar(covar,var,prefix=NULL)
}
