% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlotCat2.R
\name{separateEq}
\alias{separateEq}
\title{Separate equation}
\usage{
separateEq(equation)
}
\arguments{
\item{equation}{string. Equations to separate}
}
\description{
Separate equation
}
\examples{
equation="( a1 + b1 * W )"
separateEq(equation)
}
