% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModel.R
\name{setPositionNodes}
\alias{setPositionNodes}
\title{Set Position of nodes}
\usage{
setPositionNodes(
  nodes,
  arrows,
  radx = 0.08,
  rady = 0.06,
  xmargin = 0.02,
  ymargin = 0.02,
  xlim = c(-0.3, 1.35),
  ylim = c(-0.07, 1.05),
  parallel2 = FALSE,
  parallel3 = FALSE
)
}
\arguments{
\item{nodes}{A data.frame of nodes}

\item{arrows}{A data.frame of arrows}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{xmargin}{horizontal margin between nodes}

\item{ymargin}{vertical margin between nodes}

\item{xlim}{the x limits (min,max) of the plot}

\item{ylim}{the y limits (min,max) of the plot}

\item{parallel2}{logical}

\item{parallel3}{logical}
}
\description{
Set Position of nodes
}
