% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{standardizeDf}
\alias{standardizeDf}
\title{standardize data}
\usage{
standardizeDf(df, names)
}
\arguments{
\item{df}{A data.frame}

\item{names}{column names to mean centering}
}
\description{
standardize data
}
