% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{str_detect2}
\alias{str_detect2}
\title{Extension of str_detect to list}
\usage{
str_detect2(list, pattern)
}
\arguments{
\item{list}{a list}

\item{pattern}{pattern to look for}
}
\description{
Extension of str_detect to list
}
\examples{
site=list(c("a","c"),c("a","b","c"))
str_detect2(site,"b")
}
