% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{str_setdiff}
\alias{str_setdiff}
\title{Remove matched pattern from string}
\usage{
str_setdiff(string = "a,c", pattern = "a")
}
\arguments{
\item{string}{string}

\item{pattern}{pattern to look for}
}
\description{
Remove matched pattern from string
}
