% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{strGrouping}
\alias{strGrouping}
\title{Make Grouping equation}
\usage{
strGrouping(x, groupby = "X")
}
\arguments{
\item{x}{character vector}

\item{groupby}{name of groupby}
}
\description{
Make Grouping equation
}
