% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationBK.R
\name{summary.mediationBK}
\alias{summary.mediationBK}
\title{S3 method for class mediationBK}
\usage{
\method{summary}{mediationBK}(object, ...)
}
\arguments{
\item{object}{An object of class mediationBK}

\item{...}{Further arguments to be passed to summary()}
}
\description{
S3 method for class mediationBK
}
