% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{teams}
\alias{teams}
\title{Teams data set}
\format{A data.frame with 60 rows and 4 variables
\describe{
   \item{dysfunc}{Dysfunctional team behavior}
   \item{negtone}{Negative affective tone}
   \item{negexp}{Negative expressivity}
   \item{perform}{Team performance}
}}
\source{
Cole, M. S., Walter, F., & Bruch, H. (2008). Affective mechanisms linking dysfunctional behavior to performance in work teams: A moderated mediation study. Journal of Applied Psychology, 93, 945-958.

\url{http://www.afhayes.com/introduction-to-mediation-moderation-and-conditional-process-analysis.html}
}
\usage{
teams
}
\description{
Teams data set
}
\keyword{datasets}
