% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{theme_clean2}
\alias{theme_clean2}
\title{Clean theme for ggCor}
\usage{
theme_clean2(base_size = 12, xangle = 45, yangle = 0)
}
\arguments{
\item{base_size}{base font size}

\item{xangle}{x-axis text angle}

\item{yangle}{y-axis text angle}
}
\description{
Clean theme for ggCor
}
