% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleMediation.R
\name{treatInteraction}
\alias{treatInteraction}
\title{unfold interaction}
\usage{
treatInteraction(var)
}
\arguments{
\item{var}{name of variables}
}
\description{
unfold interaction
}
\examples{
var="X*M"
treatInteraction(var)
var="X*M*W"
treatInteraction(var)
}
