% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{tripleInteraction}
\alias{tripleInteraction}
\title{Make triple interaction equation}
\usage{
tripleInteraction(vars, prefix = "c", suffix = 0, mode = 0, addPrefix = TRUE)
}
\arguments{
\item{vars}{variable names to be interact}

\item{prefix}{A character}

\item{suffix}{A number}

\item{mode}{A number}

\item{addPrefix}{A logical}
}
\description{
Make triple interaction equation
}
\examples{
vars=c("negemot","sex","age")
tripleInteraction(vars)
tripleInteraction(vars,mode=1)
}
