% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlotCat2.R
\name{unfold}
\alias{unfold}
\title{Unfold equations}
\usage{
unfold(string, var = "W", mode = -1)
}
\arguments{
\item{string}{Character vectors with equation}

\item{var}{name of variable}

\item{mode}{integer. Default value is -1. If 0, get intercept, If 1, get slope}
}
\description{
Unfold equations
}
\examples{
string=c("(a1+b1*W)*(a2+b2*W)*(a3+b3*W)","a1+b1*W")
unfold(string)
}
