% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModel2.R
\name{vars2df}
\alias{vars2df}
\title{Make data.frame from a list of vars}
\usage{
vars2df(vars, mpos = c(0.5, 0.9))
}
\arguments{
\item{vars}{A list}

\item{mpos}{A numeric vector of length 2}
}
\description{
Make data.frame from a list of vars
}
\examples{
vars=list(name=list(c("tenure","age")),site=list(c("a","b")))
vars2df(vars)
vars=list(name=list(c("milk","hair")),matrix=list(c(1,0,0,0,0,0,1,0,0,0)),pos=5)
vars2df(vars)
}
