% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{vif}
\alias{vif}
\title{Variance Inflation Factors
Calculates variance-inflation and generalized variance-inflation factors for linear, generalized linear, and other models.}
\usage{
vif(mod)
}
\arguments{
\item{mod}{an object that responds to coef, vcov, and model.matrix, such as an lm or glm object.}
}
\value{
A vector of vifs, or a matrix containing one row for each term in the model, and columns for the GVIF, df.
}
\description{
Variance Inflation Factors
Calculates variance-inflation and generalized variance-inflation factors for linear, generalized linear, and other models.
}
