% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMC.R
\name{ztable.compareMC}
\alias{ztable.compareMC}
\title{S3 method for class compareMC}
\usage{
\method{ztable}{compareMC}(x, digits = digits, ...)
}
\arguments{
\item{x}{An object of class compareMC}

\item{digits}{integer indicating the number of decimal places}

\item{...}{further arguments to be passed to ztable}
}
\description{
S3 method for class compareMC
}
\examples{
\donttest{
require(ztable)
fit=lm(govact~negemot*age,data=glbwarm)
res=compareMC(fit)
ztable(res)
}
}
