% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catInteraction.R
\name{addCatVar}
\alias{addCatVar}
\title{Add dummy vars to data.frame}
\usage{
addCatVar(df, varname)
}
\arguments{
\item{df}{A data.frame}

\item{varname}{Variable name to be converted as factor and add dummies}
}
\description{
Add dummy vars to data.frame
}
\examples{
addCatVar(mtcars,"cyl")
}
