% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCatVars.R
\name{addCatVars}
\alias{addCatVars}
\title{Add dummy vars to data.frame}
\usage{
addCatVars(df, varnames, groupLetter = "d", mode = 1)
}
\arguments{
\item{df}{A data.frame}

\item{varnames}{Variable name to be converted as factor and add dummies}

\item{groupLetter}{A character}

\item{mode}{if mode is 2, apply different coding system}
}
\description{
Add dummy vars to data.frame
}
\examples{
addCatVars(mtcars,c("cyl","carb"))
protest1=addCatVars(protest,"protest")
head(protest1)
}
