% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catInteraction.R
\name{catInteraction}
\alias{catInteraction}
\title{Make interaction equation with dummy categorical variable"}
\usage{
catInteraction(Y = "liking", X = NULL, W = "sexism", data = NULL,
  count = NULL, prefix = "b", covar = list(), corr = TRUE)
}
\arguments{
\item{Y}{Name of dependent variable}

\item{X}{Optional.Name of independent variable}

\item{W}{Name of moderator variable}

\item{data}{A data.frame}

\item{count}{length of unique values of independent variable}

\item{prefix}{A prefix}

\item{covar}{A list}

\item{corr}{A logical}
}
\description{
Make interaction equation with dummy categorical variable"
}
\examples{
cat(catInteraction(Y="mpg",W="wt",count=3))
cat(catInteraction(Y="mpg",X="cyl",W="wt",data=mtcars))
}
