% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{deleteSingleNumber}
\alias{deleteSingleNumber}
\title{remove coefficent number of equation}
\usage{
deleteSingleNumber(equation)
}
\arguments{
\item{equation}{string}
}
\description{
remove coefficent number of equation
}
