% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModel.R
\name{drawModel}
\alias{drawModel}
\title{Draw statistical diagram with an object of class lavaan}
\usage{
drawModel(semfit, labels = NULL, nodelabels = NULL,
  whatLabel = "name", mode = 1, nodemode = 1, xmargin = 0.02,
  radx = NULL, ymargin = 0.02, xlim = c(-0.2, 1.2), ylim = xlim,
  rady = 0.06, maxypos = 0.6, minypos = 0, ypos = c(1, 0.5),
  mpos = c(0.5, 0.9), xinterval = NULL, yinterval = NULL,
  xspace = NULL, label.pos = list(), digits = 3)
}
\arguments{
\item{semfit}{AN object of class lavaan}

\item{labels}{list of variable names}

\item{nodelabels}{list of nodes names}

\item{whatLabel}{What should the edge labels indicate in the path diagram? Choices are c("est","name")}

\item{mode}{integer If 1, models with categorical X}

\item{nodemode}{integer If 1, separate node name and node label}

\item{xmargin}{horizontal margin between nodes}

\item{radx}{horizontal radius of the box.}

\item{ymargin}{vertical margin between nodes}

\item{xlim}{the x limits (min,max) of the plot}

\item{ylim}{the y limits (min,max) of the plot}

\item{rady}{vertical radius of the box.}

\item{maxypos}{maximal y position of X or W variables}

\item{minypos}{minimal y position of X or W variables}

\item{ypos}{The x and y position of Y node. Default value is c(1,0.5)}

\item{mpos}{The x and y position of M node. Default value is c(0.5,0.9)}

\item{xinterval}{numeric. Horizontal intervals among labels for nodes and nodes}

\item{yinterval}{numeric. Vertical intervals among labels for nodes and nodes}

\item{xspace}{numeric. Horizontal distance bewteen nodes}

\item{label.pos}{Optional list of arrow label position}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Draw statistical diagram with an object of class lavaan
}
\examples{
library(lavaan)
labels=list(X="protest",W="sexism",M="respappr",Y="liking")
moderator=list(name="sexism",site=list(c("a","c")))
data1=addCatVars(protest,"protest",mode=3)
model=catMediation(X="protest",M="respappr",Y="liking",moderator=moderator,data=protest,maxylev=6)
semfit=sem(model,data=data1)
nodelabels=list(D1="Ind.Protest",D2="Col.Protest",W="sexism",M="respappr",Y="liking")
drawModel(semfit,labels=labels,nodelabels=nodelabels,whatLabel="name",xlim=c(-0.4,1.3),xinterval=0.26)
drawModel(semfit,labels=labels)
labels=list(X="cyl",M=c("am","hp"),Y="mpg",W="vs")
moderator=list(name=c("vs"),site=list(c("a","b")))
model=multipleMediation(labels=labels,moderator=moderator,data=mtcars)
semfit=sem(model=model,data=mtcars)
drawModel(semfit,labels=labels)
}
