% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2table.R
\name{eq2var}
\alias{eq2var}
\title{make data.frame with equation}
\usage{
eq2var(eq, labels = list())
}
\arguments{
\item{eq}{equation}

\item{labels}{A list}
}
\description{
make data.frame with equation
}
\examples{
labels=list(X="frame",M="justify",Y="donate",W="skeptic")
eq="donate~justify+frame+skeptic+frame:skeptic"
eq2var(eq,labels=labels)
eq="Y~M+W+M:W+X+X:W"
labels=list(X="X",M="M",Y="Y")
eq2var(eq,labels=labels)
eq="wine~baby+milk+baby:milk"
labels=list(X="baby",M=c("wine","tent","sand"),Y="tile",W="milk")
eq2var(eq,labels=labels)
eq="M3~X"
eq="Y~M1+M2+X"
labels=list(X="X",M=c("M1","M2"),Y="Y")
eq2var(eq,labels=labels)
}
