% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{extractRange}
\alias{extractRange}
\title{Extract range from a data.frame}
\usage{
extractRange(res, mod, what = "indirect")
}
\arguments{
\item{res}{A data.frame}

\item{mod}{Name of moderator}

\item{what}{string}
}
\description{
Extract range from a data.frame
}
