% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{findNames}
\alias{findNames}
\title{convert a vector of names with list}
\usage{
findNames(labels, nodeslabels = list(), names, exact = FALSE)
}
\arguments{
\item{labels}{A named list}

\item{nodeslabels}{A named list}

\item{names}{A character vector to look for}

\item{exact}{A logical}
}
\description{
convert a vector of names with list
}
\examples{
labels=list(X="wt",Mi="am",Y="mpg");names=c("X","MiX","Y")
findNames(labels,names=names)
}
