% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPlotCat.R
\name{getCatSlopeDf}
\alias{getCatSlopeDf}
\title{Make data summarizing regression slopes and intercepts}
\usage{
getCatSlopeDf(labels = NULL, data, add.label = FALSE, maxylev = 6)
}
\arguments{
\item{labels}{Named list of variables}

\item{data}{A data.frame}

\item{add.label}{logical}

\item{maxylev}{maximal unique length of categorical variable}
}
\description{
Make data summarizing regression slopes and intercepts
}
\examples{
labels=list(X="protest",W="sexism",Y="liking")
getCatSlopeDf(labels=labels,data=protest)
}
