% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSummary.R
\name{getYhat1}
\alias{getYhat1}
\title{Get Yhat value from simple mediation}
\usage{
getYhat1(
  data,
  X = NULL,
  M = NULL,
  Y = NULL,
  labels,
  digits = 3,
  maxylev = 6,
  mode = 1
)
}
\arguments{
\item{data}{A data.frame}

\item{X}{Name of independant variable}

\item{M}{Name of moderator variable}

\item{Y}{Name of dependant variable}

\item{labels}{optional list of labels}

\item{digits}{Integer indicating the number of decimal places}

\item{maxylev}{maximal unique length of categorical variable}

\item{mode}{Numeric. One of 1:4. 1= simple indicator coding, 2= sequential coding, 3= Helmert coding, 4= effect coding}
}
\description{
Get Yhat value from simple mediation
}
\examples{
data=protest
labels=list(X="protest",M="respappr",Y="liking")
getYhat1(data=protest,labels=labels)
}
