% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMediation.R
\name{makeIndirectEquationCat}
\alias{makeIndirectEquationCat}
\title{Make indirect equation for categorical variables}
\usage{
makeIndirectEquationCat(
  X,
  M,
  temp1,
  temp2,
  temp3,
  moderatorNames,
  range = TRUE,
  data = NULL,
  rangemode = 1,
  probs = c(0.16, 0.5, 0.84),
  grouplabels = list()
)
}
\arguments{
\item{X}{A character vector}

\item{M}{A character vector}

\item{temp1}{A character vector}

\item{temp2}{A character vector}

\item{temp3}{A character vector}

\item{moderatorNames}{A character vector}

\item{range}{A logical}

\item{data}{A data.frame}

\item{rangemode}{range mode}

\item{probs}{numeric vector of probabilities with values in [0,1]}

\item{grouplabels}{A list}
}
\description{
Make indirect equation for categorical variables
}
