% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{matrix2df}
\alias{matrix2df}
\title{Make data.frame with matrix}
\usage{
matrix2df(matrix = c(1, 1, 1, 0, 0, 1))
}
\arguments{
\item{matrix}{a numeric vector}
}
\description{
Make data.frame with matrix
}
\examples{
matrix2df(c(1,1,1,0,0,1))
}
