% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{medSummaryTable}
\alias{medSummaryTable}
\title{Make a table summarizing the mediation effects}
\usage{
medSummaryTable(x, vanilla = TRUE, ...)
}
\arguments{
\item{x}{An object of class medSummary or medSummary2 or lavaan}

\item{vanilla}{A logical}

\item{...}{Further arguments to be passed to medSummary}
}
\description{
Make a table summarizing the mediation effects
}
