% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{medSummaryTable1}
\alias{medSummaryTable1}
\title{Make a table summarizing the mediation effects}
\usage{
medSummaryTable1(x, vanilla = TRUE, showP = FALSE)
}
\arguments{
\item{x}{An object of class medSummary}

\item{vanilla}{A logical}

\item{showP}{A logical}
}
\description{
Make a table summarizing the mediation effects
}
