% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{modelsSummary}
\alias{modelsSummary}
\title{Make Summary for Model Coefficients}
\usage{
modelsSummary(fit = NULL, labels = NULL, vars = NULL,
  moderator = NULL, data = NULL, prefix = "b", constant = "iy",
  autoPrefix = TRUE)
}
\arguments{
\item{fit}{A list of objects of class lm}

\item{labels}{optional list}

\item{vars}{optional list}

\item{moderator}{optional list}

\item{data}{optional data.frame}

\item{prefix}{A character}

\item{constant}{A string vector}

\item{autoPrefix}{logical automatic numbering of prefix}
}
\value{
A data.frame
}
\description{
Make Summary for Model Coefficients
}
\examples{
fit1=lm(mpg~wt,data=mtcars)
fit2=lm(mpg~wt*hp*am,data=mtcars)
fit=list(fit1,fit2)
labels=list(Y="mpg",X="wt",W="hp",Z="am")
modelsSummary(fit,labels=labels)
fit1=lm(withdraw~estress+ese+sex+tenure,data=estress)
fit2=lm(affect~estress+ese+sex+tenure,data=estress)
fit3=lm(withdraw~estress+affect+ese+sex+tenure,data=estress)
labels=list(Y="withdraw",M="affect",X="estress",C1="ese",C2="sex",C3="tenure")
fit=list(fit1,fit2,fit3)
modelsSummary(fit,labels=labels)
labels=list(X="dysfunc",M="negtone",W="negexp",Y="perform")
moderator=list(name="negexp",site=list(c("a","b","c")))
eq=tripleEquation(labels=labels,moderator=moderator,data=teams,mode=1)
fit=eq2fit(eq,data=teams)
modelsSummary(fit,labels=labels)
labels=list(X="cond",M="pmi",Y="reaction")
modelsSummary(labels=labels,data=pmi)
}
