% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{modmedSummary}
\alias{modmedSummary}
\title{Summarize the moderated mediation}
\usage{
modmedSummary(fit, mod = "skeptic", values = NULL)
}
\arguments{
\item{fit}{An object of class lavaan}

\item{mod}{name of moderator}

\item{values}{Optional. Numeric vector}
}
\value{
A data.frame and an object of class modmedSummary
}
\description{
Summarize the moderated mediation
}
