% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedSummary.R
\name{modmedSummaryTable}
\alias{modmedSummaryTable}
\title{Make a table summarizing the moderated mediation}
\usage{
modmedSummaryTable(x, vanilla = TRUE, showP = FALSE, ...)
}
\arguments{
\item{x}{An object of class modmedSummary or class lavaan}

\item{vanilla}{A logical}

\item{showP}{logical}

\item{...}{Further arguments to be passed to modmedSummary}
}
\description{
Make a table summarizing the moderated mediation
}
