% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSummary.R
\name{modSummary}
\alias{modSummary}
\title{Make moderation effect summary}
\usage{
modSummary(fit, mod = "skeptic", values = NULL,
  boot.ci.type = "bca.simple")
}
\arguments{
\item{fit}{An object of class lavaan}

\item{mod}{name of moderatot variable}

\item{values}{optional values of moderator}

\item{boot.ci.type}{Type of bootstrapping interval. Choices are c("norm","basic","perc",bca.simple")}
}
\description{
Make moderation effect summary
}
\examples{
require(lavaan)
labels=list(X="frame",W="skeptic",Y="justify")
moderator=list(name='skeptic',site=list("c"))
model=tripleEquation(labels=labels,moderator=moderator,data=disaster,rangemode=2)
semfit=sem(model=model,data=disaster,se="boot",bootstrap=10)
\donttest{
modSummary(semfit,mod="skeptic")
modSummaryTable(semfit,mod="skeptic")
}
}
