% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{myarrow}
\alias{myarrow}
\title{Draw arrow}
\usage{
myarrow(
  from,
  to,
  lwd = 1,
  adjust = 1,
  label = "",
  label.pos = 0.5,
  arr.pos = NULL,
  radx = 0.1,
  rady = 0.06,
  xadj = NULL,
  yadj = NULL,
  curve = 0,
  dd = 0,
  ...
)
}
\arguments{
\item{from}{coordinates (x,y) of the point *from* which to draw arrow.}

\item{to}{coordinates (x,y) of the point *to* which to draw arrow.}

\item{lwd}{line width}

\item{adjust}{adjust position}

\item{label}{label}

\item{label.pos}{label position}

\item{arr.pos}{arrow position}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{xadj}{numeric x adjustment}

\item{yadj}{numeric y adjustment}

\item{curve}{integer relative size of curve (fraction of points distance)}

\item{dd}{ength of segment arm, directed away from endpoints.}

\item{...}{Further argument to be passed to straightarrow()}
}
\description{
Draw arrow
}
