% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{numberSubscript}
\alias{numberSubscript}
\title{Make number subscript}
\usage{
numberSubscript(ft, label, vanilla)
}
\arguments{
\item{ft}{An object of class flextable}

\item{label}{string vector}

\item{vanilla}{logical}
}
\description{
Make number subscript
}
