% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTx.R
\name{r2pptx}
\alias{r2pptx}
\title{Make powerpoint presentation from R file}
\usage{
r2pptx(
  file,
  filename = "report.pptx",
  keyword = c("Concept", "Diagram", "Model", "Plot", "plot", "Table", "summary"),
  rmdRemove = TRUE
)
}
\arguments{
\item{file}{source file name}

\item{filename}{destination file name}

\item{keyword}{A string vector}

\item{rmdRemove}{A logical}
}
\description{
Make powerpoint presentation from R file
}
