% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{statisticalDiagram}
\alias{statisticalDiagram}
\title{Draw statistical diagram}
\usage{
statisticalDiagram(no = 1, radx = 0.1, rady = 0.04, xmargin = 0.01,
  arrowlabel = TRUE, arrowslabels = NULL, arrowslty = NULL,
  labels = list(), whatLabel = "name", fit = NULL,
  estimateTable = NULL, digits = 3, covar = list(),
  addCovar = TRUE, includeLatentVars = FALSE, addprime = TRUE)
}
\arguments{
\item{no}{process macro model number}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{xmargin}{horizontal margin of plot}

\item{arrowlabel}{logical whether or not draw arrowlabel}

\item{arrowslabels}{A character vector}

\item{arrowslty}{linetype of arrows}

\item{labels}{A character list}

\item{whatLabel}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name")}

\item{fit}{An object of class lavaan. Result of lavaan::sem()}

\item{estimateTable}{A data.frame}

\item{digits}{Integer indicating the number of decimal places}

\item{covar}{Optional list of covariates}

\item{addCovar}{Logical. Whether or not include covariates}

\item{includeLatentVars}{A logical}

\item{addprime}{logical. Whether or not add prime to label "c"}
}
\description{
Draw statistical diagram
}
\examples{
statisticalDiagram(no=1)
covar=list(name=c("posemot","ideology","sex"),site=list(c("Y"),c("Y"),c("Y")))
statisticalDiagram(no=1,covar=covar)
covar=list(name=c("posemot","ideology","sex"),site=list(c("Mi","Y"),c("Mi","Y"),c("Mi","Y")))
statisticalDiagram(no=4,covar=covar)
statisticalDiagram(no=8,covar=covar)
}
