% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.modelSummary.R
\name{ztable.modelSummary}
\alias{ztable.modelSummary}
\title{S3 method for class 'modelSummary'}
\usage{
\method{ztable}{modelSummary}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An object of class modelSummary}

\item{digits}{integer indicating the number of decimal places}

\item{...}{Further argument to be passed to ztable}
}
\description{
S3 method for class 'modelSummary'
}
